refactor(resolved-ir): partition handwritten resolver responsibilities
