+++ submod/resolved_ir/src/ptx_resolved_ir.cpp
@@ -1,17 +1,8 @@
-#include <algorithm>
-#include <array>
-#include <bit>
-#include <cctype>
-#include <charconv>
-#include <cmath>
 #include <limits>
 #include <ptx_frontend/resolved_ir/ptx_resolved_ir.hpp>
-#include <unordered_set>
-#include <ptx_frontend/base/ptx_integer.hpp>
-#include <ptx_frontend/semantic/ptx_declaration_semantics.hpp>
 #include <string_view>
-#include <vector>
-#include "resolved_value_domains.gen.hpp"
+#include "ptx_resolved_ir_layout.hpp"
+#include "ptx_resolved_ir_private.hpp"
 
 namespace ptx_frontend::resolved_ir::check_end {
 
@@ -42,7 +33,6 @@ int32_t SyntaxVariantDescriptor::get_required_modifier_num() const {
 };  // namespace ptx_frontend::resolved_ir::check_end
 
 namespace ptx_frontend::resolved_ir {
-namespace {
 
 using check_end::OperandPresence;
 using check_end::OperandSyntaxShape;
@@ -60,3131 +50,16 @@ using check_end::SyntaxOperandLayoutDescriptor;
 using check_end::SyntaxOperandSlotDescriptor;
 using check_end::SyntaxVariantDescriptor;
 
-/**
- * @brief Check if the actual operand syntax shape is allowed by the allowed shape.
- * 
- * @param allowed allowed operand syntax shape 
- * @param actual actual operand syntax shape
- * @return true if the actual shape is allowed by the allowed shape
- * @return false otherwise
- */
-bool allows_shape(OperandSyntaxShape allowed, OperandSyntaxShape actual) {
-  using Underlying = std::underlying_type_t<OperandSyntaxShape>;
-  return (static_cast<Underlying>(allowed) & static_cast<Underlying>(actual)) !=
-         0;
-}
-
-bool allows_shape(checker::OperandShape allowed, checker::OperandShape actual) {
-  using Underlying = std::underlying_type_t<checker::OperandShape>;
-  return (static_cast<Underlying>(allowed) & static_cast<Underlying>(actual)) !=
-         0;
-}
-
-OperandSyntaxShape vector_element_syntax_shape(
-    const syntax_ast::AstVectorElement& element) {
-  return std::visit(
-      [](const auto& value) {
-        using Value = std::remove_cvref_t<decltype(value)>;
-        if constexpr (std::same_as<Value, syntax_ast::AstIdentifierRef>)
-          return OperandSyntaxShape::Identifier;
-        else
-          return OperandSyntaxShape::Immediate;
-      },
-      element);
-}
-
-bool matches_operand_slot(const SyntaxOperandSlotDescriptor& slot,
-                          const syntax_ast::AstOperand& operand) {
-  if (!allows_shape(slot.allowed_shapes,
-                    check_end::get_operand_syntax_shape(operand))) {
-    return false;
-  }
-
-  const auto* vector = std::get_if<syntax_ast::AstVectorPack>(&operand);
-  if (vector == nullptr)
-    return true;
-  if (slot.minimum_elements != 0 &&
-      (vector->elements.size() < slot.minimum_elements ||
-       vector->elements.size() > slot.maximum_elements)) {
-    return false;
-  }
-  if (slot.allowed_element_shapes == OperandSyntaxShape{})
-    return true;
-  return std::ranges::all_of(vector->elements, [&slot](const auto& element) {
-    return allows_shape(slot.allowed_element_shapes,
-                        vector_element_syntax_shape(element));
-  });
-}
-
-std::optional<ResolveDiagnostic> diagnose_modern_pack_mismatch(
-    const SyntaxOperandLayoutDescriptor& layout,
-    const syntax_ast::AstInstruction& ast) {
-  if (ast.operands.size() != layout.slots.size())
-    return std::nullopt;
-
-  for (size_t index = 0; index < layout.slots.size(); ++index) {
-    const auto& slot = layout.slots[index];
-    const auto* vector =
-        std::get_if<syntax_ast::AstVectorPack>(&ast.operands[index]);
-    if (slot.minimum_elements == 0 || vector == nullptr ||
-        !allows_shape(slot.allowed_shapes,
-                      check_end::get_operand_syntax_shape(ast.operands[index]))) {
-      continue;
-    }
-    bool other_slots_match = true;
-    for (size_t other_index = 0; other_index < layout.slots.size();
-         ++other_index) {
-      if (other_index != index &&
-          !matches_operand_slot(layout.slots[other_index],
-                                ast.operands[other_index])) {
-        other_slots_match = false;
-        break;
-      }
-    }
-    if (!other_slots_match)
-      continue;
-    if (vector->elements.size() < slot.minimum_elements ||
-        vector->elements.size() > slot.maximum_elements) {
-      return ResolveDiagnostic{
-          .range = vector->range,
-          .message = fmt::format("Vector operand requires {} to {} elements.",
-                                 slot.minimum_elements, slot.maximum_elements),
-      };
-    }
-    for (const auto& element : vector->elements) {
-      if (allows_shape(slot.allowed_element_shapes,
-                       vector_element_syntax_shape(element))) {
-        continue;
-      }
-      const auto range = std::visit(
-          [](const auto& value) { return value.syntax.range; }, element);
-      return ResolveDiagnostic{
-          .range = range,
-          .message =
-              "Vector operand element has a shape not accepted by this "
-              "instruction layout.",
-      };
-    }
-  }
-  return std::nullopt;
-}
-
-/**
- * @brief Check if the actual operands of an instruction match the operand layout descriptor.
- * 
- * @param layout operand layout descriptor
- * @param ast syntax AST instruction
- * @return true if the actual operands match the layout descriptor
- * @return false otherwise
- */
-bool matches_operand_layout(const SyntaxOperandLayoutDescriptor& layout,
-                            const syntax_ast::AstInstruction& ast) {
-  if (layout.kind == check_end::OperandLayoutKind::Call ||
-      layout.kind == check_end::OperandLayoutKind::IndirectCall) {
-    if (ast.operands.size() != layout.slots.size())
-      return false;
-    for (size_t index = 0; index < layout.slots.size(); ++index) {
-      if (!matches_operand_slot(layout.slots[index], ast.operands[index])) {
-        return false;
-      }
-    }
-    const bool indirect =
-        layout.kind == check_end::OperandLayoutKind::IndirectCall;
-    const size_t minimum_operands = indirect ? 2 : 1;
-    if (ast.operands.size() == minimum_operands)
-      return true;
-    if (ast.operands.size() == minimum_operands + 1) {
-      const auto* arguments =
-          std::get_if<syntax_ast::AstCallParameterList>(&ast.operands[1]);
-      return arguments != nullptr &&
-             arguments->kind == syntax_ast::AstCallParameterListKind::Input;
-    }
-    if (ast.operands.size() == minimum_operands + 2) {
-      const auto* returns =
-          std::get_if<syntax_ast::AstCallParameterList>(&ast.operands[0]);
-      const auto* arguments =
-          std::get_if<syntax_ast::AstCallParameterList>(&ast.operands[2]);
-      return returns != nullptr && arguments != nullptr &&
-             returns->kind == syntax_ast::AstCallParameterListKind::Return &&
-             arguments->kind == syntax_ast::AstCallParameterListKind::Input;
-    }
-    return false;
-  }
-  if (ast.operands.size() > layout.slots.size())
-    return false;
-
-  for (size_t index = 0; index < layout.slots.size(); ++index) {
-    const SyntaxOperandSlotDescriptor& slot = layout.slots[index];
-    if (index == ast.operands.size()) {
-      if (slot.presence == OperandPresence::Required)
-        return false;
-      continue;
-    }
-
-    if (!matches_operand_slot(slot, ast.operands[index])) {
-      return false;
-    }
-  }
-  return true;
-}
-
-/**
- * Return whether ``candidate`` accepts a strict subset of ``other`` syntax.
- *
- * Availability is deliberately not considered here: resolve has no target
- * context. This makes an ``imm`` compatibility layout win over a later
- * ``reg_or_imm`` layout for the same immediate syntax, while still rejecting
- * layouts that are equally specific or incomparable.
- */
-bool is_more_specific_operand_layout(
-    const SyntaxOperandLayoutDescriptor& candidate,
-    const SyntaxOperandLayoutDescriptor& other) {
-  if (candidate.slots.size() != other.slots.size())
-    return false;
-
-  using Underlying = std::underlying_type_t<OperandSyntaxShape>;
-  bool strictly_more_specific = false;
-  for (size_t index = 0; index < candidate.slots.size(); ++index) {
-    const auto& candidate_slot = candidate.slots[index];
-    const auto& other_slot = other.slots[index];
-    if (candidate_slot.presence == OperandPresence::Optional &&
-        other_slot.presence == OperandPresence::Required)
-      return false;
-    strictly_more_specific |= candidate_slot.presence != other_slot.presence;
-
-    const auto candidate_shapes =
-        static_cast<Underlying>(candidate_slot.allowed_shapes);
-    const auto other_shapes =
-        static_cast<Underlying>(other_slot.allowed_shapes);
-    if ((candidate_shapes & other_shapes) != candidate_shapes)
-      return false;
-    strictly_more_specific |= candidate_shapes != other_shapes;
-
-    if (candidate_slot.minimum_elements == 0) {
-      if (other_slot.minimum_elements != 0)
-        return false;
-      continue;
-    }
-
-    if (other_slot.minimum_elements != 0) {
-      if (candidate_slot.minimum_elements < other_slot.minimum_elements ||
-          candidate_slot.maximum_elements > other_slot.maximum_elements) {
-        return false;
-      }
-      strictly_more_specific |=
-          candidate_slot.minimum_elements != other_slot.minimum_elements ||
-          candidate_slot.maximum_elements != other_slot.maximum_elements;
-    } else {
-      strictly_more_specific = true;
-    }
-
-    if (other_slot.allowed_element_shapes == OperandSyntaxShape{})
-      continue;
-    if (candidate_slot.allowed_element_shapes == OperandSyntaxShape{})
-      return false;
-
-    const auto candidate_element_shapes =
-        static_cast<Underlying>(candidate_slot.allowed_element_shapes);
-    const auto other_element_shapes =
-        static_cast<Underlying>(other_slot.allowed_element_shapes);
-    if ((candidate_element_shapes & other_element_shapes) !=
-        candidate_element_shapes) {
-      return false;
-    }
-    strictly_more_specific |= candidate_element_shapes != other_element_shapes;
-  }
-  return strictly_more_specific;
-}
-
-/**
- * @brief Select the operand layout descriptor that matches the actual operands of an instruction.
- * 
- * @param variant instruction variant descriptor
- * @param ast syntax AST instruction
- * @return the selected syntax layout and its index
- */
-struct SelectedOperandLayout {
-  const SyntaxOperandLayoutDescriptor& descriptor;
-  size_t index;
-};
-
-std::expected<SelectedOperandLayout, ResolveDiagnostic> select_operand_layout(
-    const SyntaxVariantDescriptor& variant,
-    const syntax_ast::AstInstruction& ast) {
-  std::vector<SelectedOperandLayout> matches;
-  for (size_t index = 0; index < variant.operand_layouts.size(); ++index) {
-    const auto& layout = variant.operand_layouts[index];
-    if (!matches_operand_layout(layout, ast))
-      continue;
-    matches.push_back(
-        SelectedOperandLayout{.descriptor = layout, .index = index});
-  }
-
-  if (matches.size() == 1)
-    return matches.front();
-  if (matches.size() > 1) {
-    for (const auto& candidate : matches) {
-      const bool more_specific_than_all = std::ranges::all_of(
-          matches, [&candidate](const SelectedOperandLayout& other) {
-            return candidate.index == other.index ||
-                   is_more_specific_operand_layout(candidate.descriptor,
-                                                   other.descriptor);
-          });
-      if (more_specific_than_all)
-        return candidate;
-    }
-    throw ResolveException(fmt::format(
-        "Descriptor variant '{}': multiple operand layouts match one syntax "
-        "instruction without a unique most-specific layout.",
-        variant.variant_name));
-  }
-
-  for (const auto& layout : variant.operand_layouts) {
-    if (const auto diagnostic = diagnose_modern_pack_mismatch(layout, ast))
-      return std::unexpected(*diagnostic);
-  }
-
-  return std::unexpected(ResolveDiagnostic{
-      .range = ast.range,
-      .message = fmt::format(
-          "Operands do not match any layout of instruction variant '{}'.",
-          variant.variant_name),
-  });
-}
-
-/**
- * @brief Find the variant descriptor of an instruction by its name.
- * 
- * @param instruction instruction descriptor
- * @param name variant name string
- * @return const SyntaxVariantDescriptor&
- */
-const SyntaxVariantDescriptor& find_syntax_variant_descriptor(
-    const SyntaxInstructionDescriptor& instruction, std::string_view name) {
-  const auto& descriptors = instruction.variants;
-  const auto it = std::ranges::find_if(
-      descriptors, [name](const SyntaxVariantDescriptor& descriptor) {
-        return descriptor.variant_name == name;
-      });
-  if (it == descriptors.end()) {
-    throw ResolveException(
-        fmt::format("Descriptor for '{}' has no variant named '{}'.",
-                    instruction.Opcode_name, name));
-  }
-  return *it;
-}
-
-template <typename T, size_t N>
-constexpr std::optional<T> lookup_ptx_suffix(
-    const std::array<generated_detail::PtxSuffixEntry<T>, N>& entries,
-    std::string_view spelling) {
-  if (spelling.starts_with('.'))
-    spelling.remove_prefix(1);
-  for (const auto& entry : entries) {
-    if (entry.suffix == spelling)
-      return entry.value;
-  }
-  return std::nullopt;
-}
-
-std::optional<ScalarType> scalar_type_from_ptx_name(std::string_view spelling) {
-  return lookup_ptx_suffix(generated_detail::kScalarTypes, spelling);
-}
-
-std::expected<WithLocs<ScalarType>, ResolveDiagnostic> resolve_scalar_type(
-    const syntax_ast::AstModifier& modifier) {
-  const auto type = scalar_type_from_ptx_name(modifier.syntax.text);
-  if (!type) {
-    return std::unexpected(ResolveDiagnostic{
-        .range = modifier.syntax.range,
-        .message =
-            fmt::format("Unknown scalar type '{}'.", modifier.syntax.text),
-    });
-  }
-  return WithLocs<ScalarType>{*type, modifier.syntax.range};
-}
-
-std::optional<RoundingMode> rounding_mode_from_ptx_name(
-    std::string_view spelling) {
-  return lookup_ptx_suffix(generated_detail::kRoundingModes, spelling);
-}
-
-std::expected<WithLocs<RoundingMode>, ResolveDiagnostic> resolve_rounding_mode(
-    const syntax_ast::AstModifier& modifier) {
-  const auto mode = rounding_mode_from_ptx_name(modifier.syntax.text);
-  if (!mode) {
-    return std::unexpected(ResolveDiagnostic{
-        .range = modifier.syntax.range,
-        .message =
-            fmt::format("Unknown rounding mode '{}'.", modifier.syntax.text),
-    });
-  }
-  return WithLocs<RoundingMode>{*mode, modifier.syntax.range};
-}
-
-std::optional<ComparisonOperator> comparison_operator_from_ptx_name(
-    std::string_view spelling) {
-  return lookup_ptx_suffix(generated_detail::kComparisonOperators, spelling);
-}
-
-std::expected<WithLocs<ComparisonOperator>, ResolveDiagnostic>
-resolve_comparison_operator(const syntax_ast::AstModifier& modifier) {
-  const auto value = comparison_operator_from_ptx_name(modifier.syntax.text);
-  if (!value) {
-    return std::unexpected(ResolveDiagnostic{
-        .range = modifier.syntax.range,
-        .message = fmt::format("Unknown comparison operator '{}'.",
-                               modifier.syntax.text),
-    });
-  }
-  return WithLocs<ComparisonOperator>{*value, modifier.syntax.range};
-}
-
-std::optional<BooleanOperator> boolean_operator_from_ptx_name(
-    std::string_view spelling) {
-  return lookup_ptx_suffix(generated_detail::kBooleanOperators, spelling);
-}
-
-std::expected<WithLocs<BooleanOperator>, ResolveDiagnostic>
-resolve_boolean_operator(const syntax_ast::AstModifier& modifier) {
-  const auto value = boolean_operator_from_ptx_name(modifier.syntax.text);
-  if (!value) {
-    return std::unexpected(ResolveDiagnostic{
-        .range = modifier.syntax.range,
-        .message =
-            fmt::format("Unknown boolean operator '{}'.", modifier.syntax.text),
-    });
-  }
-  return WithLocs<BooleanOperator>{*value, modifier.syntax.range};
-}
-
-std::optional<CacheOperator> cache_operator_from_ptx_name(
-    std::string_view spelling) {
-  return lookup_ptx_suffix(generated_detail::kCacheOperators, spelling);
-}
-
-std::expected<WithLocs<CacheOperator>, ResolveDiagnostic> resolve_cache_operator(
-    const syntax_ast::AstModifier& modifier) {
-  const auto cache_operator = cache_operator_from_ptx_name(modifier.syntax.text);
-  if (!cache_operator) {
-    return std::unexpected(ResolveDiagnostic{
-        .range = modifier.syntax.range,
-        .message =
-            fmt::format("Unknown cache operator '{}'.", modifier.syntax.text),
-    });
-  }
-  return WithLocs<CacheOperator>{*cache_operator, modifier.syntax.range};
-}
-
-std::optional<EvictionPriority> eviction_priority_from_ptx_name(
-    std::string_view spelling) {
-  if (spelling.starts_with(".L1::") || spelling.starts_with(".L2::"))
-    spelling.remove_prefix(std::string_view{".L1::"}.size());
-  return lookup_ptx_suffix(generated_detail::kEvictionPriorities, spelling);
-}
-
-std::expected<WithLocs<EvictionPriority>, ResolveDiagnostic>
-resolve_eviction_priority(const syntax_ast::AstModifier& modifier) {
-  const auto value = eviction_priority_from_ptx_name(modifier.syntax.text);
-  if (!value) {
-    return std::unexpected(ResolveDiagnostic{
-        .range = modifier.syntax.range,
-        .message = fmt::format("Unknown eviction priority '{}'.",
-                               modifier.syntax.text),
-    });
-  }
-  return WithLocs<EvictionPriority>{*value, modifier.syntax.range};
-}
-
-std::optional<MemoryConsistency> memory_consistency_from_ptx_name(
-    std::string_view spelling) {
-  return lookup_ptx_suffix(generated_detail::kMemoryConsistencies, spelling);
-}
-
-std::expected<WithLocs<MemoryConsistency>, ResolveDiagnostic>
-resolve_memory_consistency(const syntax_ast::AstModifier& modifier) {
-  const auto value = memory_consistency_from_ptx_name(modifier.syntax.text);
-  if (!value) {
-    return std::unexpected(ResolveDiagnostic{
-        .range = modifier.syntax.range,
-        .message = fmt::format("Unknown memory consistency '{}'.",
-                               modifier.syntax.text),
-    });
-  }
-  return WithLocs<MemoryConsistency>{*value, modifier.syntax.range};
-}
-
-std::optional<MemoryScope> memory_scope_from_ptx_name(
-    std::string_view spelling) {
-  return lookup_ptx_suffix(generated_detail::kMemoryScopes, spelling);
-}
-
-std::expected<WithLocs<MemoryScope>, ResolveDiagnostic> resolve_memory_scope(
-    const syntax_ast::AstModifier& modifier) {
-  const auto value = memory_scope_from_ptx_name(modifier.syntax.text);
-  if (!value) {
-    return std::unexpected(ResolveDiagnostic{
-        .range = modifier.syntax.range,
-        .message = fmt::format("Unknown memory scope '{}'.", modifier.syntax.text),
-    });
-  }
-  return WithLocs<MemoryScope>{*value, modifier.syntax.range};
-}
-
-std::expected<WithLocs<MbarrierPhaseType>, ResolveDiagnostic>
-resolve_mbarrier_phase_type(const syntax_ast::AstModifier& modifier) {
-  const auto value = lookup_ptx_suffix(
-      generated_detail::kMbarrierPhaseTypes, modifier.syntax.text);
-  if (!value)
-    return std::unexpected(ResolveDiagnostic{
-        .range = modifier.syntax.range,
-        .message = fmt::format("Unknown mbarrier phase type '{}'.",
-                               modifier.syntax.text),
-    });
-  return WithLocs<MbarrierPhaseType>{*value, modifier.syntax.range};
-}
-
-std::expected<WithLocs<MbarrierLayout>, ResolveDiagnostic>
-resolve_mbarrier_layout(const syntax_ast::AstModifier& modifier) {
-  const auto value = lookup_ptx_suffix(generated_detail::kMbarrierLayouts,
-                                       modifier.syntax.text);
-  if (!value)
-    return std::unexpected(ResolveDiagnostic{
-        .range = modifier.syntax.range,
-        .message = fmt::format("Unknown mbarrier layout '{}'.",
-                               modifier.syntax.text),
-    });
-  return WithLocs<MbarrierLayout>{*value, modifier.syntax.range};
-}
-
-std::expected<WithLocs<AsyncProxyKind>, ResolveDiagnostic>
-resolve_async_proxy_kind(const syntax_ast::AstModifier& modifier) {
-  const auto value = lookup_ptx_suffix(generated_detail::kAsyncProxyKinds,
-                                       modifier.syntax.text);
-  if (!value)
-    return std::unexpected(ResolveDiagnostic{
-        .range = modifier.syntax.range,
-        .message = fmt::format("Unknown async proxy kind '{}'.",
-                               modifier.syntax.text),
-    });
-  return WithLocs<AsyncProxyKind>{*value, modifier.syntax.range};
-}
-
-std::expected<WithLocs<ProxyKindPair>, ResolveDiagnostic>
-resolve_proxy_kind_pair(const syntax_ast::AstModifier& modifier) {
-  const auto value = lookup_ptx_suffix(generated_detail::kProxyKindPairs,
-                                       modifier.syntax.text);
-  if (!value)
-    return std::unexpected(ResolveDiagnostic{
-        .range = modifier.syntax.range,
-        .message = fmt::format("Unknown proxy kind pair '{}'.",
-                               modifier.syntax.text),
-    });
-  return WithLocs<ProxyKindPair>{*value, modifier.syntax.range};
-}
-
-std::optional<VectorArity> vector_arity_from_ptx_name(
-    std::string_view spelling) {
-  return lookup_ptx_suffix(generated_detail::kVectorArities, spelling);
-}
-
-std::expected<WithLocs<VectorArity>, ResolveDiagnostic> resolve_vector_arity(
-    const syntax_ast::AstModifier& modifier) {
-  const auto arity = vector_arity_from_ptx_name(modifier.syntax.text);
-  if (!arity) {
-    return std::unexpected(ResolveDiagnostic{
-        .range = modifier.syntax.range,
-        .message =
-            fmt::format("Unknown vector arity '{}'.", modifier.syntax.text),
-    });
-  }
-  return WithLocs<VectorArity>{*arity, modifier.syntax.range};
-}
-
-std::optional<MemoryStateSpace> memory_state_space_from_ptx_name(
-    std::string_view spelling) {
-  return lookup_ptx_suffix(generated_detail::kMemoryStateSpaces, spelling);
-}
-
-std::expected<WithLocs<MemoryStateSpace>, ResolveDiagnostic>
-resolve_memory_state_space(const syntax_ast::AstModifier& modifier) {
-  const auto state_space =
-      memory_state_space_from_ptx_name(modifier.syntax.text);
-  if (!state_space) {
-    return std::unexpected(ResolveDiagnostic{
-        .range = modifier.syntax.range,
-        .message =
-            fmt::format("Unknown memory state space '{}'.",
-                        modifier.syntax.text),
-    });
-  }
-  return WithLocs<MemoryStateSpace>{*state_space, modifier.syntax.range};
-}
-
-ParameterAddressQualifier parameter_address_qualifier_from_modifier(
-    std::string_view spelling) noexcept {
-  if (spelling == ".param::entry")
-    return ParameterAddressQualifier::Entry;
-  if (spelling == ".param::func")
-    return ParameterAddressQualifier::Function;
-  return ParameterAddressQualifier::Default;
-}
-
-struct ParsedNumberedRegister {
-  std::string_view prefix;
-  uint32_t index;
-};
-
-std::expected<ParsedNumberedRegister, ResolveDiagnostic>
-parse_numbered_register(const syntax_ast::AstIdentifierRef& identifier,
-                        SourceRange diagnostic_range,
-                        std::string_view expected_description) {
-  const std::string_view spelling = identifier.syntax.text;
-  size_t digit_begin = spelling.size();
-  while (digit_begin > 0 &&
-         std::isdigit(static_cast<unsigned char>(spelling[digit_begin - 1]))) {
-    --digit_begin;
-  }
-  if (spelling.size() < 3 || spelling.front() != '%' ||
-      digit_begin == spelling.size() || digit_begin == 1) {
-    return std::unexpected(ResolveDiagnostic{
-        .range = diagnostic_range,
-        .message = fmt::format("Expected {}, got '{}'.", expected_description,
-                               spelling),
-    });
-  }
-
-  uint32_t index = 0;
-  const char* first = spelling.data() + digit_begin;
-  const char* last = spelling.data() + spelling.size();
-  const auto [end, error] = std::from_chars(first, last, index);
-  if (error != std::errc{} || end != last) {
-    return std::unexpected(ResolveDiagnostic{
-        .range = diagnostic_range,
-        .message =
-            fmt::format("Register '{}' has an invalid numeric ID.", spelling),
-    });
-  }
-  return ParsedNumberedRegister{
-      .prefix = spelling.substr(1, digit_begin - 1),
-      .index = index,
-  };
-}
-
-std::optional<uint32_t> numbered_register_index(std::string_view spelling) {
-  size_t digit_begin = spelling.size();
-  while (digit_begin > 0 &&
-         std::isdigit(static_cast<unsigned char>(spelling[digit_begin - 1]))) {
-    --digit_begin;
-  }
-  if (digit_begin == spelling.size())
-    return std::nullopt;
-
-  uint32_t index = 0;
-  const auto [end, error] = std::from_chars(
-      spelling.data() + digit_begin, spelling.data() + spelling.size(), index);
-  if (error != std::errc{} || end != spelling.data() + spelling.size())
-    return std::nullopt;
-  return index;
-}
-
-std::expected<ResolvedRegisterRef, ResolveDiagnostic> resolve_bound_register(
-    const syntax_ast::AstIdentifierRef& identifier,
-    ResolvedRegisterClass register_class, const ResolveContext& context,
-    SourceRange range,
-    std::optional<uint8_t> required_vector_width = std::nullopt) {
-  if (binding::isSpecialRegister(identifier.syntax.text)) {
-    return std::unexpected(ResolveDiagnostic{
-        .range = range,
-        .message = fmt::format(
-            "Special register '{}' is not supported by this resolved operand.",
-            identifier.syntax.text),
-    });
-  }
-  const auto lookup =
-      context.symbols.lookup(context.scope, identifier.syntax.text);
-  if (!lookup) {
-    return std::unexpected(ResolveDiagnostic{
-        .range = range,
-        .message =
-            fmt::format("Unresolved register '{}'.", identifier.syntax.text),
-    });
-  }
-
-  const binding::Symbol& symbol = context.symbols.symbol(lookup->symbol);
-  const bool register_valued_role =
-      symbol.kind == binding::SymbolKind::Variable ||
-      symbol.kind == binding::SymbolKind::InputParameter ||
-      symbol.kind == binding::SymbolKind::ReturnParameter;
-  if (!register_valued_role ||
-      symbol.state_space != syntax_ast::AstStateSpace::Register) {
-    return std::unexpected(ResolveDiagnostic{
-        .range = range,
-        .message = fmt::format("Symbol '{}' is not a .reg variable.",
-                               identifier.syntax.text),
-    });
-  }
-  if (!symbol.type) {
-    throw ResolveException(fmt::format(
-        "Register symbol '{}' has no declaration type.", symbol.name));
-  }
-  const auto declared_type = scalar_type_from_ptx_name(*symbol.type);
-  if (!declared_type) {
-    return std::unexpected(ResolveDiagnostic{
-        .range = symbol.declaration_range,
-        .message =
-            fmt::format("Register '{}' has unsupported declared type '{}'.",
-                        symbol.name, *symbol.type),
-    });
-  }
-
-  const bool is_predicate = *declared_type == ScalarType::Pred;
-  if (register_class == ResolvedRegisterClass::Predicate && !is_predicate) {
-    return std::unexpected(ResolveDiagnostic{
-        .range = range,
-        .message = fmt::format(
-            "Expected a predicate register, but '{}' is declared '{}'.",
-            identifier.syntax.text, *symbol.type),
-    });
-  }
-  if (register_class == ResolvedRegisterClass::General && is_predicate) {
-    return std::unexpected(ResolveDiagnostic{
-        .range = range,
-        .message = fmt::format(
-            "Expected a non-predicate register, but '{}' is declared '.pred'.",
-            identifier.syntax.text),
-    });
-  }
-  if (required_vector_width) {
-    if (symbol.vector_width != required_vector_width) {
-      return std::unexpected(ResolveDiagnostic{
-          .range = range,
-          .message =
-              fmt::format("Expected a .reg .v{} register, but '{}' is not one.",
-                          *required_vector_width, identifier.syntax.text),
-      });
-    }
-  } else if (symbol.vector_width) {
-    return std::unexpected(ResolveDiagnostic{
-        .range = range,
-        .message = fmt::format(
-            "Vector register '{}' is not supported by this resolved operand.",
-            identifier.syntax.text),
-    });
-  }
-
-  return ResolvedRegisterRef{
-      .spelling = identifier.syntax.text,
-      .register_class = register_class,
-      .index = numbered_register_index(identifier.syntax.text),
-      .symbol_id = lookup->symbol,
-      .parameterized_index = lookup->parameterized_index,
-      .declared_type = *declared_type,
-      .vector_width = symbol.vector_width,
-  };
-}
-
-std::expected<WithLocs<ResolvedRegisterRef>, ResolveDiagnostic>
-resolve_register(const syntax_ast::AstOperand& operand,
-                 const ResolveContext* context) {
-  const auto* identifier = std::get_if<syntax_ast::AstIdentifierRef>(&operand);
-  if (identifier == nullptr) {
-    return std::unexpected(ResolveDiagnostic{
-        .range = syntax_ast::sourceRange(operand),
-        .message = "Expected a register operand.",
-    });
-  }
-
-  if (context != nullptr) {
-    auto value =
-        resolve_bound_register(*identifier, ResolvedRegisterClass::General,
-                               *context, identifier->syntax.range);
-    if (!value)
-      return std::unexpected(value.error());
-    return WithLocs<ResolvedRegisterRef>{std::move(*value),
-                                         identifier->syntax.range};
-  }
-
-  const auto parsed = parse_numbered_register(
-      *identifier, identifier->syntax.range, "a numbered register");
-  if (!parsed)
-    return std::unexpected(parsed.error());
-  if (parsed->prefix == "p") {
-    return std::unexpected(ResolveDiagnostic{
-        .range = identifier->syntax.range,
-        .message = fmt::format("Expected a non-predicate register, got '{}'.",
-                               identifier->syntax.text),
-    });
-  }
-  return WithLocs<ResolvedRegisterRef>{
-      ResolvedRegisterRef{
-          .spelling = identifier->syntax.text,
-          .register_class = ResolvedRegisterClass::General,
-          .index = parsed->index,
-      },
-      identifier->syntax.range};
-}
-
-std::expected<WithLocs<ResolvedVectorRegisterRef>, ResolveDiagnostic>
-resolve_vector_register(const syntax_ast::AstOperand& operand,
-                        const ResolveContext* context) {
-  const auto* identifier = std::get_if<syntax_ast::AstIdentifierRef>(&operand);
-  if (identifier == nullptr) {
-    return std::unexpected(ResolveDiagnostic{
-        .range = syntax_ast::sourceRange(operand),
-        .message = "Expected a vector register operand.",
-    });
-  }
-  if (context == nullptr) {
-    return std::unexpected(ResolveDiagnostic{
-        .range = identifier->syntax.range,
-        .message = "A vector register operand requires a module declaration.",
-    });
-  }
-  auto value =
-      resolve_bound_register(*identifier, ResolvedRegisterClass::General,
-                             *context, identifier->syntax.range, 4);
-  if (!value)
-    return std::unexpected(value.error());
-  return WithLocs<ResolvedVectorRegisterRef>{
-      ResolvedVectorRegisterRef{.register_ref = std::move(*value)},
-      identifier->syntax.range};
-}
-
-std::expected<WithLocs<ResolvedPredicate>, ResolveDiagnostic>
-resolve_predicate_identifier(const syntax_ast::AstIdentifierRef& identifier,
-                             bool negated, SourceRange range,
-                             const ResolveContext* context) {
-  ResolvedRegisterRef register_ref{
-      .register_class = ResolvedRegisterClass::Predicate,
-  };
-  if (context != nullptr) {
-    auto value = resolve_bound_register(
-        identifier, ResolvedRegisterClass::Predicate, *context, range);
-    if (!value)
-      return std::unexpected(value.error());
-    register_ref = std::move(*value);
-  } else {
-    const auto parsed = parse_numbered_register(
-        identifier, range, "a numbered predicate register");
-    if (!parsed)
-      return std::unexpected(parsed.error());
-    if (parsed->prefix != "p") {
-      return std::unexpected(ResolveDiagnostic{
-          .range = range,
-          .message = fmt::format("Expected a predicate register, got '{}'.",
-                                 identifier.syntax.text),
-      });
-    }
-    register_ref = ResolvedRegisterRef{
-        .spelling = identifier.syntax.text,
-        .register_class = ResolvedRegisterClass::Predicate,
-        .index = parsed->index,
-    };
-  }
-
-  return WithLocs<ResolvedPredicate>{
-      ResolvedPredicate{.register_ref = std::move(register_ref),
-                        .negated = negated},
-      range};
-}
-
-std::expected<WithLocs<ResolvedPredicate>, ResolveDiagnostic> resolve_predicate(
-    const syntax_ast::AstOperand& operand, const ResolveContext* context) {
-  if (const auto* plain = std::get_if<syntax_ast::AstIdentifierRef>(&operand)) {
-    return resolve_predicate_identifier(*plain, false, plain->syntax.range,
-                                        context);
-  }
-  if (const auto* predicate =
-          std::get_if<syntax_ast::AstPredicateOperand>(&operand)) {
-    return resolve_predicate_identifier(predicate->name, predicate->negated,
-                                        predicate->range, context);
-  }
-  return std::unexpected(ResolveDiagnostic{
-      .range = syntax_ast::sourceRange(operand),
-      .message = "Expected a predicate operand.",
-  });
-}
-
-std::expected<WithLocs<ResolvedShflSyncDestination>, ResolveDiagnostic>
-resolve_shfl_destination(const syntax_ast::AstOperand& operand,
-                         bool allow_destination_sink,
-                         bool allow_predicate_sink,
-                         const ResolveContext* context) {
-  const auto* pair =
-      std::get_if<syntax_ast::AstRegisterPredicatePair>(&operand);
-  if (pair == nullptr) {
-    return std::unexpected(ResolveDiagnostic{.range = syntax_ast::sourceRange(operand),
-                                               .message = "Expected d|p destination."});
-  }
-  std::optional<WithLoc<ResolvedRegisterRef>> data;
-  if (pair->dst.syntax.text == "_") {
-    if (!allow_destination_sink) {
-      return std::unexpected(ResolveDiagnostic{
-          .range = pair->dst.syntax.range,
-          .message = "The '_' sink is not allowed in this d|p destination.",
-      });
-    }
-  } else {
-    auto resolved_data =
-        resolve_register(syntax_ast::AstOperand{pair->dst}, context);
-    if (!resolved_data)
-      return std::unexpected(resolved_data.error());
-    data = WithLoc<ResolvedRegisterRef>{std::move(resolved_data->value),
-                                        pair->dst.syntax.range};
-  }
-  std::optional<WithLoc<ResolvedPredicate>> predicate;
-  if (pair->predicate.syntax.text == "_") {
-    if (!allow_predicate_sink) {
-      return std::unexpected(ResolveDiagnostic{
-          .range = pair->predicate.syntax.range,
-          .message = "The '_' sink is not allowed as the predicate half of this d|p destination.",
-      });
-    }
-  } else {
-    auto resolved_predicate = resolve_predicate_identifier(
-        pair->predicate, false, pair->predicate.syntax.range, context);
-    if (!resolved_predicate)
-      return std::unexpected(resolved_predicate.error());
-    predicate = WithLoc<ResolvedPredicate>{
-        std::move(resolved_predicate->value), pair->predicate.syntax.range};
-  }
-  if (!data && !predicate) {
-    return std::unexpected(ResolveDiagnostic{
-        .range = pair->range,
-        .message = "A d|p destination must retain a data or predicate output.",
-    });
-  }
-  WithLocs<ResolvedShflSyncDestination> result{
-      ResolvedShflSyncDestination{.data = std::move(data),
-                                  .predicate = std::move(predicate)},
-      pair->range};
-  result.locs = {pair->dst.syntax.range, pair->predicate.syntax.range};
-  return result;
-}
-
-std::expected<WithLocs<ResolvedPredicatePair>, ResolveDiagnostic>
-resolve_predicate_pair(const syntax_ast::AstOperand& operand,
-                       const ResolveContext* context) {
-  const auto* pair =
-      std::get_if<syntax_ast::AstRegisterPredicatePair>(&operand);
-  if (pair == nullptr) {
-    return std::unexpected(ResolveDiagnostic{
-        .range = syntax_ast::sourceRange(operand),
-        .message = "Expected a predicate-pair operand.",
-    });
-  }
-  auto first = resolve_predicate_identifier(pair->dst, false,
-                                            pair->dst.syntax.range, context);
-  if (!first)
-    return std::unexpected(first.error());
-  auto second = resolve_predicate_identifier(
-      pair->predicate, false, pair->predicate.syntax.range, context);
-  if (!second)
-    return std::unexpected(second.error());
-  WithLocs<ResolvedPredicatePair> result{
-      ResolvedPredicatePair{.first = std::move(first->value),
-                            .second = std::move(second->value)},
-      pair->range};
-  result.locs = {pair->dst.syntax.range, pair->predicate.syntax.range};
-  return result;
-}
-
-std::expected<WithLocs<ResolvedBranchTarget>, ResolveDiagnostic>
-resolve_branch_target(const syntax_ast::AstOperand& operand,
-                      const ResolveContext* context) {
-  const auto* target = std::get_if<syntax_ast::AstBranchTarget>(&operand);
-  if (target == nullptr) {
-    return std::unexpected(ResolveDiagnostic{
-        .range = syntax_ast::sourceRange(operand),
-        .message = "Expected a direct branch target.",
-    });
-  }
-
-  ResolvedBranchTarget resolved{.spelling = target->name.syntax.text};
-  if (context != nullptr) {
-    const binding::ScopeId function_scope =
-        context->function_scope.value_or(context->scope);
-    const auto lookup =
-        context->symbols.lookup(context->scope, target->name.syntax.text);
-    if (!lookup) {
-      return std::unexpected(ResolveDiagnostic{
-          .range = target->range,
-          .message = fmt::format("Unresolved branch target '{}'.",
-                                 target->name.syntax.text),
-      });
-    }
-    const binding::Symbol& symbol = context->symbols.symbol(lookup->symbol);
-    if (symbol.kind != binding::SymbolKind::Label ||
-        symbol.scope != function_scope) {
-      return std::unexpected(ResolveDiagnostic{
-          .range = target->range,
-          .message = fmt::format(
-              "Branch target '{}' must name a label in the current function.",
-              target->name.syntax.text),
-      });
-    }
-    resolved.symbol_id = lookup->symbol;
-  }
-
-  return WithLocs<ResolvedBranchTarget>{std::move(resolved), target->range};
-}
-
-std::expected<WithLocs<ResolvedBranchTargetSet>, ResolveDiagnostic>
-resolve_branch_target_set(const syntax_ast::AstOperand& operand,
-                          const ResolveContext* context) {
-  const auto* target_set =
-      std::get_if<syntax_ast::AstBranchTargetSet>(&operand);
-  if (target_set == nullptr) {
-    return std::unexpected(ResolveDiagnostic{
-        .range = syntax_ast::sourceRange(operand),
-        .message = "Expected a brx.idx branch target list.",
-    });
-  }
-
-  ResolvedBranchTargetSet resolved{.spelling = target_set->name.syntax.text};
-  if (context != nullptr) {
-    const binding::ScopeId function_scope =
-        context->function_scope.value_or(context->scope);
-    const auto lookup =
-        context->symbols.lookup(context->scope, target_set->name.syntax.text);
-    if (!lookup) {
-      return std::unexpected(ResolveDiagnostic{
-          .range = target_set->range,
-          .message = fmt::format("Unresolved brx.idx branch target list '{}'.",
-                                 target_set->name.syntax.text),
-      });
-    }
-    const binding::Symbol& symbol = context->symbols.symbol(lookup->symbol);
-    if (symbol.kind != binding::SymbolKind::BranchTargetSet ||
-        symbol.scope != function_scope) {
-      return std::unexpected(ResolveDiagnostic{
-          .range = target_set->range,
-          .message = fmt::format(
-              "brx.idx branch target list '{}' must name a function-local "
-              ".branchtargets declaration.",
-              target_set->name.syntax.text),
-      });
-    }
-    resolved.symbol_id = symbol.id;
-  }
-  return WithLocs<ResolvedBranchTargetSet>{std::move(resolved),
-                                            target_set->range};
-}
-
-std::expected<ResolvedCallParameterRef, ResolveDiagnostic>
-resolve_call_parameter(const syntax_ast::AstIdentifierRef& identifier,
-                       const ResolveContext* context) {
-  ResolvedCallParameterRef resolved{.spelling = identifier.syntax.text};
-  if (context == nullptr)
-    return resolved;
-
-  const auto lookup =
-      context->symbols.lookup(context->scope, identifier.syntax.text);
-  if (!lookup) {
-    return std::unexpected(ResolveDiagnostic{
-        .range = identifier.syntax.range,
-        .message = fmt::format("Unresolved call parameter '{}'.",
-                               identifier.syntax.text),
-    });
-  }
-  const binding::Symbol& symbol = context->symbols.symbol(lookup->symbol);
-  const bool parameter_or_variable =
-      symbol.kind == binding::SymbolKind::Variable ||
-      symbol.kind == binding::SymbolKind::InputParameter ||
-      symbol.kind == binding::SymbolKind::ReturnParameter ||
-      symbol.kind == binding::SymbolKind::CallParameter;
-  const bool allowed_space = symbol.state_space &&
-      (*symbol.state_space == syntax_ast::AstStateSpace::Register ||
-       *symbol.state_space == syntax_ast::AstStateSpace::Parameter);
-  if (!parameter_or_variable || !allowed_space) {
-    return std::unexpected(ResolveDiagnostic{
-        .range = identifier.syntax.range,
-        .message = fmt::format("Call parameter '{}' must name a .reg or "
-                               ".param variable.",
-                               identifier.syntax.text),
-    });
-  }
-  if (*symbol.state_space == syntax_ast::AstStateSpace::Parameter &&
-      (symbol.kind == binding::SymbolKind::InputParameter ||
-       symbol.kind == binding::SymbolKind::ReturnParameter)) {
-    return std::unexpected(ResolveDiagnostic{
-        .range = identifier.syntax.range,
-        .message = fmt::format(
-            "Formal .param parameter '{}' cannot be used directly as a call "
-            "argument; use a register or body-local .param object.",
-            identifier.syntax.text),
-    });
-  }
-  resolved.symbol_id = lookup->symbol;
-  resolved.parameterized_index = lookup->parameterized_index;
-  resolved.state_space = symbol.state_space;
-  if (!symbol.type) {
-    throw ResolveException(fmt::format(
-        "Call parameter symbol '{}' has no declaration type.", symbol.name));
-  }
-  const auto declared_type = scalar_type_from_ptx_name(*symbol.type);
-  if (!declared_type) {
-    return std::unexpected(ResolveDiagnostic{
-        .range = symbol.declaration_range,
-        .message = fmt::format(
-            "Call parameter '{}' has unsupported declared type '{}'.",
-            symbol.name, *symbol.type),
-    });
-  }
-  resolved.declared_type = *declared_type;
-  return resolved;
-}
-
-std::expected<WithLocs<ResolvedFunctionRef>, ResolveDiagnostic>
-resolve_direct_call_target(const syntax_ast::AstOperand& operand,
-                           const ResolveContext* context) {
-  const auto* target = std::get_if<syntax_ast::AstCallTarget>(&operand);
-  if (target == nullptr) {
-    return std::unexpected(ResolveDiagnostic{
-        .range = syntax_ast::sourceRange(operand),
-        .message = "Expected a direct call target.",
-    });
-  }
-
-  ResolvedFunctionRef resolved{.spelling = target->name.syntax.text};
-  if (context != nullptr) {
-    const auto lookup =
-        context->symbols.lookup(context->scope, target->name.syntax.text);
-    if (!lookup) {
-      return std::unexpected(ResolveDiagnostic{
-          .range = target->range,
-          .message = fmt::format("Unresolved call target '{}'.",
-                                 target->name.syntax.text),
-      });
-    }
-    const binding::Symbol& symbol = context->symbols.symbol(lookup->symbol);
-    if (symbol.kind != binding::SymbolKind::Function) {
-      const bool is_register = symbol.kind == binding::SymbolKind::Variable &&
-          symbol.state_space == syntax_ast::AstStateSpace::Register;
-      return std::unexpected(ResolveDiagnostic{
-          .range = target->range,
-          .message = is_register
-              ? "Indirect call register targets require a function-local "
-                ".callprototype or .calltargets metadata operand."
-              : fmt::format("Call target '{}' must name a function.",
-                            target->name.syntax.text),
-      });
-    }
-    if (symbol.function_is_entry) {
-      return std::unexpected(ResolveDiagnostic{
-          .range = target->range,
-          .message = fmt::format(
-              "Direct call target '{}' must name a device .func, not an .entry.",
-              target->name.syntax.text),
-      });
-    }
-    resolved.symbol_id = symbol.id;
-    resolved.is_entry = symbol.function_is_entry;
-  }
-  return WithLocs<ResolvedFunctionRef>{std::move(resolved), target->range};
-}
-
-std::expected<WithLocs<ResolvedIndirectCallee>, ResolveDiagnostic>
-resolve_indirect_callee(const syntax_ast::AstOperand& operand,
-                        const ResolveContext* context) {
-  if (const auto* target = std::get_if<syntax_ast::AstCallTarget>(&operand)) {
-    auto register_ref = resolve_register(syntax_ast::AstOperand{target->name},
-                                         context);
-    if (!register_ref)
-      return std::unexpected(register_ref.error());
-    return WithLocs<ResolvedIndirectCallee>{
-        ResolvedIndirectCallee{std::move(register_ref->value)}, target->range};
-  }
-
-  const auto* metadata = std::get_if<syntax_ast::AstCallTargetSet>(&operand);
-  if (metadata == nullptr) {
-    return std::unexpected(ResolveDiagnostic{
-        .range = syntax_ast::sourceRange(operand),
-        .message = "Expected an indirect call register target or metadata label.",
-    });
-  }
-
-  ResolvedIndirectMetadataRef resolved{.spelling = metadata->name.syntax.text};
-  if (context != nullptr) {
-    const binding::ScopeId function_scope =
-        context->function_scope.value_or(context->scope);
-    const auto lookup =
-        context->symbols.lookup(context->scope, metadata->name.syntax.text);
-    if (!lookup) {
-      return std::unexpected(ResolveDiagnostic{
-          .range = metadata->range,
-          .message = fmt::format("Unresolved indirect call metadata '{}'.",
-                                 metadata->name.syntax.text),
-      });
-    }
-    const binding::Symbol& symbol = context->symbols.symbol(lookup->symbol);
-    if (lookup->parameterized_index ||
-        symbol.kind == binding::SymbolKind::Variable) {
-      return std::unexpected(ResolveDiagnostic{
-          .range = metadata->range,
-          .message = "Indirect call metadata variables and call-table arrays "
-                     "are not supported.",
-      });
-    }
-    if (symbol.scope != function_scope ||
-        (symbol.kind != binding::SymbolKind::CallPrototype &&
-         symbol.kind != binding::SymbolKind::CallTargetSet)) {
-      return std::unexpected(ResolveDiagnostic{
-          .range = metadata->range,
-          .message = fmt::format(
-              "Indirect call metadata '{}' must name a function-local "
-              ".callprototype or .calltargets declaration.",
-              metadata->name.syntax.text),
-      });
-    }
-    resolved.symbol_id = symbol.id;
-    resolved.declaration_kind = symbol.kind;
-  }
-  return WithLocs<ResolvedIndirectCallee>{
-      ResolvedIndirectCallee{std::move(resolved)}, metadata->range};
-}
-
-std::expected<WithLocs<ResolvedCallParameterRef>, ResolveDiagnostic>
-resolve_call_return_parameter(const syntax_ast::AstOperand& operand,
-                              const ResolveContext* context) {
-  const auto* group = std::get_if<syntax_ast::AstCallParameterList>(&operand);
-  if (group == nullptr ||
-      group->kind != syntax_ast::AstCallParameterListKind::Return ||
-      group->parameters.size() != 1) {
-    return std::unexpected(ResolveDiagnostic{
-        .range = syntax_ast::sourceRange(operand),
-        .message = "Expected a single call return parameter.",
-    });
-  }
-  const auto* identifier =
-      std::get_if<syntax_ast::AstIdentifierRef>(&group->parameters.front());
-  if (identifier == nullptr) {
-    return std::unexpected(ResolveDiagnostic{
-        .range = group->range,
-        .message = "A call return parameter must be a .reg or .param variable.",
-    });
-  }
-  auto resolved = resolve_call_parameter(*identifier, context);
-  if (!resolved)
-    return std::unexpected(resolved.error());
-  return WithLocs<ResolvedCallParameterRef>{std::move(*resolved),
-                                             group->range};
-}
-
-std::expected<WithLocs<ResolvedCallArguments>, ResolveDiagnostic>
-resolve_call_arguments(const syntax_ast::AstOperand& operand,
-                       const ResolveContext* context) {
-  const auto* group = std::get_if<syntax_ast::AstCallParameterList>(&operand);
-  if (group == nullptr ||
-      group->kind != syntax_ast::AstCallParameterListKind::Input) {
-    return std::unexpected(ResolveDiagnostic{
-        .range = syntax_ast::sourceRange(operand),
-        .message = "Expected a call input parameter group.",
-    });
-  }
-
-  ResolvedCallArguments resolved;
-  resolved.values.reserve(group->parameters.size());
-  for (const auto& parameter : group->parameters) {
-    if (const auto* identifier =
-            std::get_if<syntax_ast::AstIdentifierRef>(&parameter)) {
-      auto value = resolve_call_parameter(*identifier, context);
-      if (!value)
-        return std::unexpected(value.error());
-      resolved.values.emplace_back(ResolvedCallArgument{std::move(*value)},
-                                   identifier->syntax.range);
-      continue;
-    }
-    const auto& immediate = std::get<syntax_ast::AstImmediate>(parameter);
-    resolved.values.emplace_back(
-        ResolvedCallArgument{ResolvedCallLiteral{
-            .spelling = immediate.syntax.text, .kind = immediate.kind}},
-        immediate.syntax.range);
-  }
-  return WithLocs<ResolvedCallArguments>{std::move(resolved), group->range};
-}
-
-std::expected<WithLocs<ResolvedSpecialRegisterRef>, ResolveDiagnostic>
-resolve_special_register(const syntax_ast::AstOperand& operand) {
-  std::string spelling;
-  SourceRange range;
-  std::optional<base::VectorComponent> component;
-  if (const auto* identifier =
-          std::get_if<syntax_ast::AstIdentifierRef>(&operand)) {
-    spelling = identifier->syntax.text;
-    range = identifier->syntax.range;
-  } else if (const auto* member =
-                 std::get_if<syntax_ast::AstVectorMember>(&operand)) {
-    spelling = member->base.syntax.text + member->selector.text;
-    range = member->range;
-    switch (member->selector.text.back()) {
-      case 'x':
-        component = base::VectorComponent::X;
-        break;
-      case 'y':
-        component = base::VectorComponent::Y;
-        break;
-      case 'z':
-        component = base::VectorComponent::Z;
-        break;
-      default:
-        break;
-    }
-  } else {
-    return std::unexpected(ResolveDiagnostic{
-        .range = syntax_ast::sourceRange(operand),
-        .message = "Expected a special-register operand.",
-    });
-  }
-
-  auto info = base::lookup(spelling);
-  if (!info) {
-    return std::unexpected(ResolveDiagnostic{
-        .range = range,
-        .message = fmt::format("Unknown special register '{}'.", spelling),
-    });
-  }
-  if (info->vector_width != 1) {
-    return std::unexpected(ResolveDiagnostic{
-        .range = range,
-        .message = fmt::format(
-            "Special register '{}' is a vector; select a scalar component.",
-            spelling),
-    });
-  }
-  return WithLocs<ResolvedSpecialRegisterRef>{
-      ResolvedSpecialRegisterRef{.spelling = std::move(spelling),
-                                 .id = info->id,
-                                 .component = component},
-      range};
-}
-
-std::expected<WithLocs<ResolvedPredicateSource>, ResolveDiagnostic>
-resolve_predicate_source(const syntax_ast::AstOperand& operand,
-                         const ResolveContext* context) {
-  const auto range = syntax_ast::sourceRange(operand);
-  if (std::holds_alternative<syntax_ast::AstIdentifierRef>(operand) ||
-      std::holds_alternative<syntax_ast::AstVectorMember>(operand)) {
-    auto special = resolve_special_register(operand);
-    if (special) {
-      if (base::metadata(special->value.id).element_type !=
-          base::ScalarType::Pred) {
-        return std::unexpected(ResolveDiagnostic{
-            .range = range,
-            .message = fmt::format("Expected a predicate special register, got '{}'.",
-                                   special->value.spelling),
-        });
-      }
-      return WithLocs<ResolvedPredicateSource>{
-          ResolvedPredicateSource{std::move(special->value)}, range};
-    }
-    const auto* identifier =
-        std::get_if<syntax_ast::AstIdentifierRef>(&operand);
-    if (identifier == nullptr || base::lookup(identifier->syntax.text))
-      return std::unexpected(special.error());
-  }
-  auto predicate = resolve_predicate(operand, context);
-  if (!predicate)
-    return std::unexpected(predicate.error());
-  return WithLocs<ResolvedPredicateSource>{
-      ResolvedPredicateSource{std::move(predicate->value)}, range};
-}
-
-std::expected<WithLocs<ResolvedVectorSpecialRegisterRef>, ResolveDiagnostic>
-resolve_vector_special_register(const syntax_ast::AstOperand& operand) {
-  const auto* identifier = std::get_if<syntax_ast::AstIdentifierRef>(&operand);
-  if (identifier == nullptr) {
-    return std::unexpected(ResolveDiagnostic{
-        .range = syntax_ast::sourceRange(operand),
-        .message = "Expected a vector special-register base operand.",
-    });
-  }
-  const auto info = base::lookup(identifier->syntax.text);
-  if (!info || info->vector_width != 4) {
-    return std::unexpected(ResolveDiagnostic{
-        .range = identifier->syntax.range,
-        .message = fmt::format("Expected a four-component special register, got '{}'.",
-                               identifier->syntax.text),
-    });
-  }
-  return WithLocs<ResolvedVectorSpecialRegisterRef>{
-      ResolvedVectorSpecialRegisterRef{.spelling = identifier->syntax.text,
-                                       .id = info->id},
-      identifier->syntax.range};
-}
-
-std::expected<ResolvedImmediate, ResolveDiagnostic> resolve_immediate_value(
-    const syntax_ast::AstImmediate& immediate, ScalarType type,
-    bool require_target_range = false);
-
-/** Select whether an address consumer imposes memory-offset source limits. */
-enum class AddressImmediateDomain : uint8_t { General, MemoryOperand };
-
-/**
- * Resolve an address offset while retaining its signed-64-bit IR magnitude.
- *
- * Bracketed memory addresses additionally constrain the effective signed
- * offset, after combining the address operator with a lexical literal sign,
- * to the PTX signed-32-bit source domain.  Other consumers retain their
- * wider relocation-addend contract.
- */
-std::expected<std::optional<ResolvedAddressOffset>, ResolveDiagnostic>
-resolve_address_offset(const syntax_ast::AstAddress& address,
-                       AddressImmediateDomain domain =
-                           AddressImmediateDomain::General) {
-  if (!address.offset)
-    return std::nullopt;
-
-  const bool memory_operand = domain == AddressImmediateDomain::MemoryOperand;
-  auto value = resolve_immediate_value(address.offset->magnitude,
-                                       ScalarType::S64, true);
-  if (!value)
-    return std::unexpected(value.error());
-  const bool subtract = address.offset->operation ==
-                        syntax_ast::AstAddressOffset::Operator::Subtract;
-  if (memory_operand) {
-    const uint64_t source_bits = value->integer_source_bits.value_or(value->bits);
-    const uint64_t magnitude =
-        value->is_negative ? uint64_t{0} - source_bits : source_bits;
-    const bool effective_negative = subtract != value->is_negative;
-    const uint64_t maximum_magnitude =
-        effective_negative ? uint64_t{1} << 31 : (uint64_t{1} << 31) - 1;
-    if (magnitude > maximum_magnitude) {
-      return std::unexpected(ResolveDiagnostic{
-          .range = address.offset->magnitude.syntax.range,
-          .message = fmt::format(
-              "Address offset magnitude '{}' is outside the signed 32-bit "
-              "range for its '{}' operator.",
-              address.offset->magnitude.syntax.text,
-              effective_negative ? "-" : "+"),
-      });
-    }
-    // ResolvedAddressOffset retains the spelling's operation separately, so
-    // retain its magnitude representation while validating the signed PTX domain.
-  }
-  return ResolvedAddressOffset{
-      .operation = subtract ? ResolvedAddressOffsetOperator::Subtract
-                            : ResolvedAddressOffsetOperator::Add,
-      .value = std::move(*value),
-  };
-}
-
-/** Return whether a declared register can hold a PTX address value. */
-bool is_address_register_type(ScalarType type) {
-  const auto kind = scalar_kind(type);
-  return (kind == base::ScalarKind::Unsigned ||
-          kind == base::ScalarKind::Signed || kind == base::ScalarKind::Bit) &&
-         scalar_size_of(type) <= sizeof(uint64_t);
-}
-
-/** Reject a bound address base whose declaration cannot represent an address. */
-std::expected<void, ResolveDiagnostic> check_address_register_type(
-    const ResolvedRegisterRef& register_ref, SourceRange range) {
-  if (register_ref.declared_type &&
-      is_address_register_type(*register_ref.declared_type))
-    return {};
-  return std::unexpected(ResolveDiagnostic{
-      .range = range,
-      .message = fmt::format(
-          "Address register '{}' has invalid declared type '{}'; expected an "
-          "integer or bit-size type no wider than 64 bits.",
-          register_ref.spelling,
-          register_ref.declared_type ? to_string(*register_ref.declared_type)
-                                     : "unknown"),
-  });
-}
-
-enum class FormalParameterAddressPolicy : uint8_t {
-  Reject,
-  PreserveParameterSpace,
-  MaterializeDeviceParameter,
-};
-
-bool is_addressable_data_symbol(const binding::Symbol& symbol,
-                                FormalParameterAddressPolicy parameter_policy) {
-  if (symbol.kind == binding::SymbolKind::Variable) {
-    return symbol.state_space == syntax_ast::AstStateSpace::Local ||
-           symbol.state_space == syntax_ast::AstStateSpace::Shared ||
-           symbol.state_space == syntax_ast::AstStateSpace::Global ||
-           symbol.state_space == syntax_ast::AstStateSpace::Constant;
-  }
-  return parameter_policy != FormalParameterAddressPolicy::Reject &&
-         (symbol.kind == binding::SymbolKind::InputParameter ||
-          symbol.kind == binding::SymbolKind::ReturnParameter ||
-          (parameter_policy ==
-               FormalParameterAddressPolicy::PreserveParameterSpace &&
-           symbol.kind == binding::SymbolKind::CallParameter)) &&
-         symbol.state_space == syntax_ast::AstStateSpace::Parameter;
-}
-
-std::expected<ResolvedSymbolRef, ResolveDiagnostic> resolve_data_symbol(
-    const syntax_ast::AstIdentifierRef& identifier,
-    const ResolveContext* context,
-    FormalParameterAddressPolicy parameter_policy) {
-  ResolvedSymbolRef resolved{.spelling = identifier.syntax.text};
-  if (context == nullptr)
-    return resolved;
-
-  if (binding::isSpecialRegister(identifier.syntax.text)) {
-    return std::unexpected(ResolveDiagnostic{
-        .range = identifier.syntax.range,
-        .message = fmt::format("Special register '{}' is not a data symbol.",
-                               identifier.syntax.text),
-    });
-  }
-  const auto lookup =
-      context->symbols.lookup(context->scope, identifier.syntax.text);
-  if (!lookup) {
-    return std::unexpected(ResolveDiagnostic{
-        .range = identifier.syntax.range,
-        .message =
-            fmt::format("Unresolved data symbol '{}'.", identifier.syntax.text),
-    });
-  }
-  const binding::Symbol& symbol = context->symbols.symbol(lookup->symbol);
-  if (!is_addressable_data_symbol(symbol, parameter_policy)) {
-    return std::unexpected(ResolveDiagnostic{
-        .range = identifier.syntax.range,
-        .message = fmt::format("Symbol '{}' is not an addressable data symbol.",
-                               identifier.syntax.text),
-    });
-  }
-
-  resolved.symbol_id = lookup->symbol;
-  resolved.parameterized_index = lookup->parameterized_index;
-  resolved.declaration_kind = symbol.kind;
-  resolved.declaration_state_space = symbol.state_space;
-  resolved.address_state_space = symbol.state_space;
-  if (symbol.kind == binding::SymbolKind::InputParameter ||
-      symbol.kind == binding::SymbolKind::ReturnParameter) {
-    // A direct formal-parameter memory address stays in .param.  Only mov
-    // address-taking materializes a device-function parameter on the stack and
-    // consequently changes the produced address to .local.
-    if (parameter_policy ==
-            FormalParameterAddressPolicy::MaterializeDeviceParameter &&
-        !context->function_is_entry) {
-      resolved.address_state_space = syntax_ast::AstStateSpace::Local;
-      // Device-function parameters require PTX 2.0 and sm_20.  PTX raised the
-      // minimum for taking a return-parameter address to 6.0.
-      resolved.address_availability = checker::AvailabilityDescriptor{
-          .minimum_ptx_version =
-              symbol.kind == binding::SymbolKind::ReturnParameter
-                  ? checker::PtxVersion{6, 0}
-                  : checker::PtxVersion{2, 0},
-          .minimum_sm_version = 20,
-      };
-    }
-  }
-  if (symbol.type)
-    resolved.declared_type = scalar_type_from_ptx_name(*symbol.type);
-  resolved.address_alignment = symbol.address_alignment;
-  return resolved;
-}
-
-std::expected<WithLocs<ResolvedSymbolRef>, ResolveDiagnostic> resolve_symbol(
-    const syntax_ast::AstOperand& operand, const ResolveContext* context) {
-  const auto* identifier = std::get_if<syntax_ast::AstIdentifierRef>(&operand);
-  if (identifier == nullptr) {
-    return std::unexpected(ResolveDiagnostic{
-        .range = syntax_ast::sourceRange(operand),
-        .message = "Expected a data-symbol operand.",
-    });
-  }
-  auto resolved = resolve_data_symbol(*identifier, context,
-                                      FormalParameterAddressPolicy::Reject);
-  if (!resolved)
-    return std::unexpected(resolved.error());
-  return WithLocs<ResolvedSymbolRef>{std::move(*resolved),
-                                     identifier->syntax.range};
-}
-
-std::expected<WithLocs<ResolvedAddress>, ResolveDiagnostic> resolve_address(
-    const syntax_ast::AstOperand& operand, const ResolveContext* context) {
-  const auto* address = std::get_if<syntax_ast::AstAddress>(&operand);
-  if (address == nullptr || !address->bracketed) {
-    return std::unexpected(ResolveDiagnostic{
-        .range = syntax_ast::sourceRange(operand),
-        .message = "Expected a bracketed address operand.",
-    });
-  }
-
-  std::optional<ResolvedAddressBase> base;
-  if (const auto* identifier =
-          std::get_if<syntax_ast::AstIdentifierRef>(&address->base)) {
-    if (context != nullptr) {
-      const auto lookup =
-          context->symbols.lookup(context->scope, identifier->syntax.text);
-      if (lookup &&
-          context->symbols.symbol(lookup->symbol).kind ==
-              binding::SymbolKind::Variable &&
-          context->symbols.symbol(lookup->symbol).state_space ==
-              syntax_ast::AstStateSpace::Register) {
-        auto register_ref =
-            resolve_bound_register(*identifier, ResolvedRegisterClass::General,
-                                   *context, identifier->syntax.range);
-        if (!register_ref)
-          return std::unexpected(register_ref.error());
-        if (auto type_check = check_address_register_type(
-                *register_ref, identifier->syntax.range);
-            !type_check)
-          return std::unexpected(type_check.error());
-        base = std::move(*register_ref);
-      } else {
-        auto symbol = resolve_data_symbol(
-            *identifier, context,
-            FormalParameterAddressPolicy::PreserveParameterSpace);
-        if (!symbol)
-          return std::unexpected(symbol.error());
-        base = std::move(*symbol);
-      }
-    } else if (identifier->syntax.text.starts_with('%')) {
-      const syntax_ast::AstOperand base_operand = *identifier;
-      auto register_ref = resolve_register(base_operand, nullptr);
-      if (!register_ref)
-        return std::unexpected(register_ref.error());
-      base = std::move(register_ref->value);
-    } else {
-      auto symbol = resolve_data_symbol(
-          *identifier, nullptr,
-          FormalParameterAddressPolicy::PreserveParameterSpace);
-      base = std::move(*symbol);
-    }
-  } else {
-    const auto& immediate = std::get<syntax_ast::AstImmediate>(address->base);
-    auto immediate_base = resolve_immediate_value(immediate, ScalarType::U32,
-                                                  true);
-    if (!immediate_base)
-      return std::unexpected(immediate_base.error());
-    if (immediate_base->is_negative) {
-      return std::unexpected(ResolveDiagnostic{
-          .range = immediate.syntax.range,
-          .message = fmt::format(
-              "Immediate address '{}' must be an unsigned 32-bit value.",
-              immediate.syntax.text),
-      });
-    }
-    base = std::move(*immediate_base);
-  }
-
-  auto offset = resolve_address_offset(*address,
-                                       AddressImmediateDomain::MemoryOperand);
-  if (!offset)
-    return std::unexpected(offset.error());
-
-  return WithLocs<ResolvedAddress>{
-      ResolvedAddress{
-          .base = std::move(*base),
-          .offset = std::move(*offset),
-          .enclosing_function_kind =
-              context == nullptr ? EnclosingFunctionKind::Unknown
-              : context->function_is_entry ? EnclosingFunctionKind::Entry
-                                           : EnclosingFunctionKind::Device,
-      },
-      address->range};
-}
-
-ResolveDiagnostic invalid_immediate(const syntax_ast::AstImmediate& immediate,
-                                    std::string message) {
-  return ResolveDiagnostic{
-      .range = immediate.syntax.range,
-      .message = std::move(message),
-  };
-}
-
-std::expected<uint64_t, ResolveDiagnostic> parse_unsigned_literal(
-    const syntax_ast::AstImmediate& immediate, std::string_view text,
-    int base) {
-  if (!text.empty() && (text.back() == 'u' || text.back() == 'U'))
-    text.remove_suffix(1);
-
-  uint64_t value = 0;
-  const auto [end, error] =
-      std::from_chars(text.data(), text.data() + text.size(), value, base);
-  if (text.empty() || error != std::errc{} ||
-      end != text.data() + text.size()) {
-    return std::unexpected(invalid_immediate(
-        immediate,
-        fmt::format("Invalid integer literal '{}'.", immediate.syntax.text)));
-  }
-  return value;
-}
-
-std::expected<ResolvedImmediate, ResolveDiagnostic> resolve_integer_literal(
-    const syntax_ast::AstImmediate& immediate, ScalarType type,
-    std::string_view text, bool negative, bool require_target_range = false) {
-  using base::ScalarKind;
-  const ScalarKind kind = scalar_kind(type);
-  if (kind != ScalarKind::Unsigned && kind != ScalarKind::Signed &&
-      kind != ScalarKind::Bit) {
-    return std::unexpected(invalid_immediate(
-        immediate,
-        fmt::format(
-            "Integer literal '{}' is incompatible with scalar type '{}'.",
-            immediate.syntax.text, to_string(type))));
-  }
-  const uint8_t byte_size = scalar_size_of(type);
-  if (byte_size == 0 || byte_size > sizeof(uint64_t)) {
-    return std::unexpected(invalid_immediate(
-        immediate,
-        fmt::format("Immediate type '{}' is not representable in 64 bits.",
-                    to_string(type))));
-  }
-  const uint8_t bit_width = byte_size * 8;
-  const uint64_t bit_mask = bit_width == 64
-                                ? std::numeric_limits<uint64_t>::max()
-                                : (uint64_t{1} << bit_width) - 1;
-
-  const auto magnitude = base::parseIntegerMagnitude(text);
-  if (!magnitude)
-    return std::unexpected(invalid_immediate(
-        immediate,
-        fmt::format("Invalid integer literal '{}'.", immediate.syntax.text)));
-
-  const bool source_is_unsigned =
-      text.ends_with('u') || text.ends_with('U') ||
-      *magnitude > static_cast<uint64_t>(std::numeric_limits<int64_t>::max());
-  const uint64_t source_bits = negative ? uint64_t{0} - *magnitude : *magnitude;
-  const bool source_is_negative =
-      !source_is_unsigned && std::bit_cast<int64_t>(source_bits) < 0;
-  if (require_target_range) {
-    const uint64_t positive_limit =
-        kind == base::ScalarKind::Signed
-            ? (uint64_t{1} << (bit_width - 1)) - 1
-            : bit_mask;
-    const uint64_t negative_limit =
-        kind == base::ScalarKind::Signed
-            ? uint64_t{1} << (bit_width - 1)
-            : bit_mask;
-    const bool representable =
-        source_is_negative
-            ? uint64_t{0} - source_bits <= negative_limit
-            : source_bits <= positive_limit;
-    if (!representable) {
-      return std::unexpected(invalid_immediate(
-          immediate,
-          fmt::format(
-              "Integer literal '{}' is out of range for scalar type '{}'.",
-              immediate.syntax.text, to_string(type))));
-    }
-  }
-  return ResolvedImmediate{
-      .bits = source_bits & bit_mask,
-      .type = type,
-      .is_negative = source_is_negative,
-      .integer_source_bits = source_bits,
-  };
-}
-
-/**
- * Convert IEEE binary64 bits to binary32 using round-to-nearest, ties-to-even.
- * Literal conversion is independent of instruction rounding and host floating
- * point modes. Overflow becomes signed infinity; underflow is gradual. NaNs
- * retain their sign and high payload bits and are quieted across precisions.
- */
-uint32_t narrow_float_literal_bits(uint64_t bits) {
-  const auto sign = static_cast<uint32_t>(bits >> 32) & 0x80000000U;
-  const auto exponent = static_cast<int>((bits >> 52) & 0x7ffU);
-  const uint64_t fraction = bits & 0x000fffffffffffffULL;
-  if (exponent == 0x7ff) {
-    return sign | 0x7f800000U |
-           (fraction == 0 ? 0U
-                          : static_cast<uint32_t>(fraction >> 29) | 0x00400000U);
-  }
-  // Every binary64 subnormal is smaller than half a binary32 subnormal ULP.
-  if (exponent == 0)
-    return sign;
-  const int unbiased_exponent = exponent - 1023;
-  if (unbiased_exponent > 127)
-    return sign | 0x7f800000U;
-  if (unbiased_exponent < -150)
-    return sign;
-
-  const uint64_t significand = (uint64_t{1} << 52) | fraction;
-  // Retain 24 bits for normals, or fewer for subnormals (shift is 29..53).
-  const int shift = unbiased_exponent >= -126 ? 29 : -97 - unbiased_exponent;
-  uint32_t rounded = static_cast<uint32_t>(significand >> shift);
-  const uint64_t remainder = significand & ((uint64_t{1} << shift) - 1);
-  const uint64_t halfway = uint64_t{1} << (shift - 1);
-  if (remainder > halfway || (remainder == halfway && (rounded & 1U)))
-    ++rounded;
-  if (unbiased_exponent < -126)
-    return sign | rounded;
-  // The retained implicit bit supplies one exponent unit; rounding may carry
-  // into the next exponent, including the infinity encoding at overflow.
-  return sign |
-         ((static_cast<uint32_t>(unbiased_exponent + 126) << 23) + rounded);
-}
-
-/**
- * Widen IEEE binary32 literal bits exactly without host floating arithmetic.
- * Subnormals and signed zero are preserved; NaNs are quieted with their sign
- * and payload retained in the high binary64 fraction bits.
- */
-uint64_t widen_float_literal_bits(uint32_t bits) {
-  const uint64_t sign = static_cast<uint64_t>(bits & 0x80000000U) << 32;
-  const uint32_t exponent = (bits >> 23) & 0xffU;
-  const uint32_t fraction = bits & 0x007fffffU;
-  if (exponent == 0xffU) {
-    return sign | 0x7ff0000000000000ULL |
-           (fraction == 0 ? 0ULL
-                          : (static_cast<uint64_t>(fraction) << 29) |
-                                0x0008000000000000ULL);
-  }
-  if (exponent != 0) {
-    return sign | (static_cast<uint64_t>(exponent + 896) << 52) |
-           (static_cast<uint64_t>(fraction) << 29);
-  }
-  if (fraction == 0)
-    return sign;
-  // A binary32 subnormal is fraction * 2^-149, normal in binary64.
-  const int leading_bit = std::bit_width(fraction) - 1;
-  return sign | (static_cast<uint64_t>(leading_bit + 874) << 52) |
-         ((static_cast<uint64_t>(fraction) << (52 - leading_bit)) &
-          0x000fffffffffffffULL);
-}
-
-/**
- * Decode a floating literal in its lexical precision, then convert to the
- * operand precision. Equal-width literals retain their exact payload bits.
- */
-std::expected<ResolvedImmediate, ResolveDiagnostic> resolve_float_bits_literal(
-    const syntax_ast::AstImmediate& immediate, ScalarType type,
-    std::string_view text, bool negative, uint8_t bit_width) {
-  if (negative) {
-    return std::unexpected(invalid_immediate(
-        immediate,
-        fmt::format("Floating bit-pattern literal '{}' cannot have a sign.",
-                    immediate.syntax.text)));
-  }
-  if (type != ScalarType::F32 && type != ScalarType::F64) {
-    return std::unexpected(invalid_immediate(
-        immediate,
-        fmt::format(
-            "Floating bit-pattern literal '{}' is incompatible with scalar "
-            "type '{}'.",
-            immediate.syntax.text, to_string(type))));
-  }
-
-  text.remove_prefix(2);  // 0f or 0d
-  const auto bits = parse_unsigned_literal(immediate, text, 16);
-  if (!bits)
-    return std::unexpected(bits.error());
-  if (bit_width == 32 && type == ScalarType::F64) {
-    return ResolvedImmediate{
-        .bits = widen_float_literal_bits(static_cast<uint32_t>(*bits)),
-        .type = type};
-  }
-  if (bit_width == 64 && type == ScalarType::F32) {
-    return ResolvedImmediate{.bits = narrow_float_literal_bits(*bits),
-                             .type = type};
-  }
-  return ResolvedImmediate{.bits = *bits, .type = type};
-}
-
-std::expected<ResolvedImmediate, ResolveDiagnostic>
-resolve_decimal_float_literal(const syntax_ast::AstImmediate& immediate,
-                              ScalarType type, std::string_view text) {
-  if (type != ScalarType::F32 && type != ScalarType::F64) {
-    return std::unexpected(invalid_immediate(
-        immediate,
-        fmt::format("Decimal floating literal '{}' is incompatible with scalar "
-                    "type '{}'.",
-                    immediate.syntax.text, to_string(type))));
-  }
-
-  if (!text.empty() && text.front() == '+') {
-    text.remove_prefix(1);
-    if (text.empty() || text.front() == '+' || text.front() == '-') {
-      return std::unexpected(invalid_immediate(
-          immediate, fmt::format("Invalid decimal floating literal '{}'.",
-                                 immediate.syntax.text)));
-    }
-  }
-
-  double value = 0.0;
-  const auto [end, error] =
-      std::from_chars(text.data(), text.data() + text.size(), value,
-                      std::chars_format::general);
-  if (text.empty() || error != std::errc{} ||
-      end != text.data() + text.size() || !std::isfinite(value)) {
-    return std::unexpected(invalid_immediate(
-        immediate, fmt::format("Invalid decimal floating literal '{}'.",
-                               immediate.syntax.text)));
-  }
-
-  if (type == ScalarType::F32) {
-    return ResolvedImmediate{
-        .bits = narrow_float_literal_bits(std::bit_cast<uint64_t>(value)),
-        .type = type};
-  }
-  return ResolvedImmediate{.bits = std::bit_cast<uint64_t>(value),
-                           .type = type};
-}
-
-std::expected<ResolvedImmediate, ResolveDiagnostic> resolve_immediate_value(
-    const syntax_ast::AstImmediate& immediate, ScalarType type,
-    bool require_target_range) {
-  std::string_view text = immediate.syntax.text;
-  bool negative = false;
-  if (!text.empty() && (text.front() == '+' || text.front() == '-')) {
-    negative = text.front() == '-';
-    text.remove_prefix(1);
-  }
-
-  switch (immediate.kind) {
-    case syntax_ast::AstImmediateKind::DecimalInteger:
-    case syntax_ast::AstImmediateKind::HexInteger:
-      return resolve_integer_literal(immediate, type, text, negative,
-                                     require_target_range);
-    case syntax_ast::AstImmediateKind::F32Hex:
-      return resolve_float_bits_literal(immediate, type, text, negative, 32);
-    case syntax_ast::AstImmediateKind::F64Hex:
-      return resolve_float_bits_literal(immediate, type, text, negative, 64);
-    case syntax_ast::AstImmediateKind::DecimalFloat:
-      return resolve_decimal_float_literal(immediate, type,
-                                           immediate.syntax.text);
-    case syntax_ast::AstImmediateKind::WarpSize:
-      // PTX defines WARP_SZ as the signed source integer constant 32.
-      return resolve_integer_literal(immediate, type, "32", negative,
-                                     require_target_range);
-  }
-  throw ResolveException("Unknown AstImmediateKind.");
-}
-
-std::expected<WithLocs<RegOrImm>, ResolveDiagnostic> resolve_reg_or_imm(
-    const syntax_ast::AstOperand& operand, ScalarType type,
-    const ResolveContext* context, bool require_target_range = false) {
-  if (const auto* identifier =
-          std::get_if<syntax_ast::AstIdentifierRef>(&operand)) {
-    auto register_ref = resolve_register(operand, context);
-    if (!register_ref)
-      return std::unexpected(register_ref.error());
-    return WithLocs<RegOrImm>{RegOrImm{register_ref->value},
-                              identifier->syntax.range};
-  }
-  if (const auto* immediate = std::get_if<syntax_ast::AstImmediate>(&operand)) {
-    auto value = resolve_immediate_value(*immediate, type,
-                                         require_target_range);
-    if (!value)
-      return std::unexpected(value.error());
-    return WithLocs<RegOrImm>{RegOrImm{*value}, immediate->syntax.range};
-  }
-  return std::unexpected(ResolveDiagnostic{
-      .range = syntax_ast::sourceRange(operand),
-      .message = "Expected a register or immediate operand.",
-  });
-}
-
-std::expected<ScalarType, ResolveDiagnostic> type_for_operand(
-    const ResolvedOperandBindingDescriptor& binding,
-    const ResolvedInstructionFields& fields, const SourceRange& range);
-
-std::expected<WithLocs<ResolvedRegisterVector>, ResolveDiagnostic>
-resolve_reg_vector(const syntax_ast::AstOperand& operand,
-                   ScalarType instruction_type,
-                   std::span<const uint8_t> allowed_arities,
-                   std::optional<uint8_t> required_arity,
-                   checker::VectorTypePolicy vector_type_policy,
-                   base::ScalarTypeSizePolicy register_width_policy,
-                   bool allow_sink,
-                   size_t sink_payload_bits,
-                   const ResolveContext* context) {
-  const auto* vector = std::get_if<syntax_ast::AstVectorPack>(&operand);
-  if (vector == nullptr) {
-    return std::unexpected(ResolveDiagnostic{
-        .range = syntax_ast::sourceRange(operand),
-        .message = "Expected a vector-pack operand.",
-    });
-  }
-  if (vector_type_policy == checker::VectorTypePolicy::Aggregate &&
-      scalar_kind(instruction_type) != base::ScalarKind::Bit) {
-    return std::unexpected(ResolveDiagnostic{
-        .range = vector->range,
-        .message = "A vector mov requires a bit-size instruction type.",
-    });
-  }
-
-  const size_t arity = vector->elements.size();
-  if (required_arity && arity != *required_arity) {
-    return std::unexpected(ResolveDiagnostic{
-        .range = vector->range,
-        .message =
-            fmt::format("This vector operand requires {} elements.",
-                        *required_arity),
-    });
-  }
-  const size_t vector_payload_bits =
-      (vector_type_policy == checker::VectorTypePolicy::Aggregate
-           ? scalar_size_of(instruction_type)
-           : arity * scalar_size_of(instruction_type)) *
-      8u;
-  if (vector_payload_bits > checker::kMaxRegisterVectorPayloadBits) {
-    return std::unexpected(ResolveDiagnostic{
-        .range = vector->range,
-        .message = fmt::format(
-            "This vector operand's payload width ({} bits) exceeds the supported "
-            "{} bit limit.",
-            vector_payload_bits,
-            checker::kMaxRegisterVectorPayloadBits),
-    });
-  }
-  if (!required_arity &&
-      std::ranges::find(allowed_arities, arity) == allowed_arities.end()) {
-    return std::unexpected(ResolveDiagnostic{
-        .range = vector->range,
-        .message = vector_type_policy == checker::VectorTypePolicy::Aggregate
-                       ? "A vector mov requires two or four elements."
-                       : "A vector operand requires two, four, or eight elements.",
-    });
-  }
-  size_t element_bytes = scalar_size_of(instruction_type);
-  if (vector_type_policy == checker::VectorTypePolicy::Aggregate) {
-    const size_t instruction_bytes = scalar_size_of(instruction_type);
-    if (instruction_bytes % arity != 0 || instruction_bytes / arity == 0) {
-      return std::unexpected(ResolveDiagnostic{
-          .range = vector->range,
-          .message = "Vector mov elements must be at least eight bits wide.",
-      });
-    }
-    element_bytes = instruction_bytes / arity;
-  }
-
-  ResolvedRegisterVector result;
-  result.elements.reserve(arity);
-  std::vector<SourceRange> locations;
-  locations.reserve(arity);
-  size_t sink_count = 0;
-  for (const auto& element : vector->elements) {
-    const auto* identifier =
-        std::get_if<syntax_ast::AstIdentifierRef>(&element);
-    if (identifier == nullptr) {
-      return std::unexpected(ResolveDiagnostic{
-          .range = std::get<syntax_ast::AstImmediate>(element).syntax.range,
-          .message = "A register-vector element must be a register or '_' sink.",
-      });
-    }
-    locations.push_back(identifier->syntax.range);
-    if (identifier->syntax.text == "_") {
-      if (!allow_sink) {
-        return std::unexpected(ResolveDiagnostic{
-            .range = identifier->syntax.range,
-            .message = "The '_' sink is allowed only in a destination vector.",
-        });
-      }
-      if (sink_payload_bits != 0 && vector_payload_bits != sink_payload_bits) {
-        return std::unexpected(ResolveDiagnostic{
-            .range = identifier->syntax.range,
-            .message = fmt::format(
-                "The '_' sink requires an exact {}-bit vector payload.",
-                sink_payload_bits),
-        });
-      }
-      ++sink_count;
-      result.elements.emplace_back(std::nullopt);
-      continue;
-    }
-
-    syntax_ast::AstOperand register_operand{*identifier};
-    auto register_ref = resolve_register(register_operand, context);
-    if (!register_ref)
-      return std::unexpected(register_ref.error());
-    if (register_ref->value.declared_type) {
-      const auto declared_type = *register_ref->value.declared_type;
-      const bool type_mismatch =
-          vector_type_policy == checker::VectorTypePolicy::Aggregate
-              ? scalar_size_of(declared_type) != element_bytes
-              : !scalar_types_compatible(declared_type, instruction_type,
-                                         register_width_policy);
-      if (type_mismatch) {
-        return std::unexpected(ResolveDiagnostic{
-            .range = identifier->syntax.range,
-            .message = fmt::format(
-                "Vector element '{}' has type '{}' incompatible with this "
-                "instruction.",
-                identifier->syntax.text, to_string(declared_type)),
-        });
-      }
-    }
-    result.elements.emplace_back(std::move(register_ref->value));
-  }
-  if (sink_count == arity) {
-    return std::unexpected(ResolveDiagnostic{
-        .range = vector->range,
-        .message = "A vector must contain at least one register.",
-    });
-  }
-  WithLocs<ResolvedRegisterVector> resolved{std::move(result)};
-  resolved.locs = std::move(locations);
-  return resolved;
-}
-
-std::expected<WithLocs<ResolvedRegisterVector>, ResolveDiagnostic>
-resolve_modern_register_vector(const syntax_ast::AstOperand& operand,
-                               const check_end::ResolvedOperandBindingDescriptor& binding,
-                               const ResolveContext* context) {
-  const auto* vector = std::get_if<syntax_ast::AstVectorPack>(&operand);
-  if (vector == nullptr) {
-    return std::unexpected(ResolveDiagnostic{
-        .range = syntax_ast::sourceRange(operand),
-        .message = "Expected a vector-pack operand.",
-    });
-  }
-  if (vector->elements.size() < binding.minimum_elements ||
-      vector->elements.size() > binding.maximum_elements) {
-    return std::unexpected(ResolveDiagnostic{
-        .range = vector->range,
-        .message = fmt::format("Vector operand requires {} to {} elements.",
-                               binding.minimum_elements,
-                               binding.maximum_elements),
-    });
-  }
-
-  ResolvedRegisterVector result;
-  result.elements.reserve(vector->elements.size());
-  std::vector<SourceRange> locations;
-  locations.reserve(vector->elements.size());
-  for (const auto& element : vector->elements) {
-    const auto* identifier = std::get_if<syntax_ast::AstIdentifierRef>(&element);
-    if (identifier == nullptr) {
-      return std::unexpected(ResolveDiagnostic{
-          .range = std::get<syntax_ast::AstImmediate>(element).syntax.range,
-          .message = "A matrix fragment element must be a register or '_' sink.",
-      });
-    }
-    locations.push_back(identifier->syntax.range);
-    if (identifier->syntax.text == "_") {
-      result.elements.emplace_back(std::nullopt);
-      continue;
-    }
-    syntax_ast::AstOperand register_operand{*identifier};
-    auto register_ref = resolve_register(register_operand, context);
-    if (!register_ref)
-      return std::unexpected(register_ref.error());
-    result.elements.emplace_back(std::move(register_ref->value));
-  }
-  WithLocs<ResolvedRegisterVector> resolved{std::move(result)};
-  resolved.locs = std::move(locations);
-  return resolved;
-}
-
-std::expected<WithLocs<ResolvedTensorCoordinate>, ResolveDiagnostic>
-resolve_tensor_coordinate(
-    const syntax_ast::AstOperand& operand,
-    const check_end::ResolvedOperandBindingDescriptor& binding,
-    const ResolvedInstructionFields& fields, const ResolveContext* context) {
-  const auto* vector = std::get_if<syntax_ast::AstVectorPack>(&operand);
-  if (vector == nullptr) {
-    return std::unexpected(ResolveDiagnostic{
-        .range = syntax_ast::sourceRange(operand),
-        .message = "Expected a vector-pack operand.",
-    });
-  }
-  if (vector->elements.size() < binding.minimum_elements ||
-      vector->elements.size() > binding.maximum_elements) {
-    return std::unexpected(ResolveDiagnostic{
-        .range = vector->range,
-        .message = fmt::format("Vector operand requires {} to {} elements.",
-                               binding.minimum_elements,
-                               binding.maximum_elements),
-    });
-  }
-
-  ResolvedTensorCoordinate result;
-  result.elements.reserve(vector->elements.size());
-  std::vector<SourceRange> locations;
-  locations.reserve(vector->elements.size());
-  std::optional<ScalarType> immediate_type;
-  for (const auto& element : vector->elements) {
-    if (const auto* identifier =
-            std::get_if<syntax_ast::AstIdentifierRef>(&element)) {
-      syntax_ast::AstOperand register_operand{*identifier};
-      auto register_ref = resolve_register(register_operand, context);
-      if (!register_ref)
-        return std::unexpected(register_ref.error());
-      locations.push_back(identifier->syntax.range);
-      result.elements.emplace_back(std::move(register_ref->value));
-      continue;
-    }
-
-    const auto& immediate = std::get<syntax_ast::AstImmediate>(element);
-    if (!immediate_type) {
-      auto type = type_for_operand(binding, fields, immediate.syntax.range);
-      if (!type)
-        return std::unexpected(type.error());
-      if (*type == ScalarType::Invalid) {
-        return std::unexpected(ResolveDiagnostic{
-            .range = immediate.syntax.range,
-            .message = "Tensor coordinate immediates require an operand scalar type.",
-        });
-      }
-      immediate_type = *type;
-    }
-    auto value = resolve_immediate_value(
-        immediate, *immediate_type,
-        binding.immediate_conversion_policy ==
-            checker::ImmediateConversionPolicy::RequireTargetRange);
-    if (!value)
-      return std::unexpected(value.error());
-    locations.push_back(immediate.syntax.range);
-    result.elements.emplace_back(std::move(*value));
-  }
-  WithLocs<ResolvedTensorCoordinate> resolved{std::move(result)};
-  resolved.locs = std::move(locations);
-  return resolved;
-}
-
-std::expected<WithLocs<ResolvedMovSource>, ResolveDiagnostic>
-resolve_mov_source(const syntax_ast::AstOperand& operand, ScalarType type,
-                   checker::OperandShape allowed_shapes,
-                   bool allow_function_symbol,
-                   const ResolveContext* context) {
-  if (type == ScalarType::B128) {
-    return std::unexpected(ResolveDiagnostic{
-        .range = syntax_ast::sourceRange(operand),
-        .message = "The .b128 mov type is available only for vector pack or "
-                   "unpack forms.",
-    });
-  }
-  const auto reject_shape =
-      [&](checker::OperandShape shape,
-          SourceRange range) -> std::optional<ResolveDiagnostic> {
-    if (allows_shape(allowed_shapes, shape))
-      return std::nullopt;
-    return ResolveDiagnostic{
-        .range = range,
-        .message =
-            "This mov variant does not accept the resolved source "
-            "operand shape.",
-    };
-  };
-  const auto reject_address_type =
-      [&](SourceRange range,
-          bool function_address = false) -> std::optional<ResolveDiagnostic> {
-    using base::ScalarKind;
-    const ScalarKind kind = scalar_kind(type);
-    const uint8_t width = scalar_size_of(type);
-    const bool integer_address =
-        kind == ScalarKind::Unsigned || kind == ScalarKind::Signed;
-    const bool data_address = integer_address || kind == ScalarKind::Bit;
-    const bool address_width = width == 4 || width == 8;
-    if (address_width && (function_address ? integer_address : data_address))
-      return std::nullopt;
-    return ResolveDiagnostic{
-        .range = range,
-        .message = function_address
-                       ? "A function address requires a 32-bit or 64-bit "
-                         "integer mov type."
-                       : "A data address requires a 32-bit or 64-bit integer "
-                         "or bit-size mov type.",
-    };
-  };
-
-  if (const auto* immediate = std::get_if<syntax_ast::AstImmediate>(&operand)) {
-    if (auto rejected = reject_shape(checker::OperandShape::Immediate,
-                                     immediate->syntax.range)) {
-      return std::unexpected(std::move(*rejected));
-    }
-    auto value = resolve_immediate_value(*immediate, type);
-    if (!value)
-      return std::unexpected(value.error());
-    return WithLocs<ResolvedMovSource>{ResolvedMovSource{std::move(*value)},
-                                       immediate->syntax.range};
-  }
-
-  if (std::holds_alternative<syntax_ast::AstVectorMember>(operand)) {
-    if (auto rejected = reject_shape(checker::OperandShape::SpecialRegister,
-                                     syntax_ast::sourceRange(operand))) {
-      return std::unexpected(std::move(*rejected));
-    }
-    auto value = resolve_special_register(operand);
-    if (!value)
-      return std::unexpected(value.error());
-    WithLocs<ResolvedMovSource> resolved{
-        ResolvedMovSource{std::move(value->value)}};
-    resolved.locs = std::move(value->locs);
-    return resolved;
-  }
-
-  if (const auto* address = std::get_if<syntax_ast::AstAddress>(&operand)) {
-    if (auto rejected =
-            reject_shape(checker::OperandShape::Address, address->range)) {
-      return std::unexpected(std::move(*rejected));
-    }
-    if (address->bracketed) {
-      return std::unexpected(ResolveDiagnostic{
-          .range = address->range,
-          .message = "Expected an unbracketed symbol-address operand.",
-      });
-    }
-    if (auto rejected = reject_address_type(address->range))
-      return std::unexpected(std::move(*rejected));
-    const auto* identifier =
-        std::get_if<syntax_ast::AstIdentifierRef>(&address->base);
-    if (identifier == nullptr) {
-      return std::unexpected(ResolveDiagnostic{
-          .range = address->range,
-          .message = "A mov address expression must use a data-symbol base.",
-      });
-    }
-    auto symbol = resolve_data_symbol(
-        *identifier, context,
-        FormalParameterAddressPolicy::MaterializeDeviceParameter);
-    if (!symbol)
-      return std::unexpected(symbol.error());
-    auto offset = resolve_address_offset(*address);
-    if (!offset)
-      return std::unexpected(offset.error());
-    ResolvedAddress value{
-        .base = std::move(*symbol),
-        .offset = std::move(*offset),
-    };
-    return WithLocs<ResolvedMovSource>{ResolvedMovSource{std::move(value)},
-                                       address->range};
-  }
-
-  const auto* identifier = std::get_if<syntax_ast::AstIdentifierRef>(&operand);
-  if (identifier == nullptr) {
-    return std::unexpected(ResolveDiagnostic{
-        .range = syntax_ast::sourceRange(operand),
-        .message = "Expected a scalar mov source operand.",
-    });
-  }
-
-  if (base::lookup(identifier->syntax.text)) {
-    if (auto rejected = reject_shape(checker::OperandShape::SpecialRegister,
-                                     identifier->syntax.range)) {
-      return std::unexpected(std::move(*rejected));
-    }
-    auto value = resolve_special_register(operand);
-    if (!value)
-      return std::unexpected(value.error());
-    WithLocs<ResolvedMovSource> resolved{
-        ResolvedMovSource{std::move(value->value)}};
-    resolved.locs = std::move(value->locs);
-    return resolved;
-  }
-
-  bool is_register = identifier->syntax.text.starts_with('%');
-  if (context != nullptr) {
-    const auto lookup =
-        context->symbols.lookup(context->scope, identifier->syntax.text);
-    if (lookup) {
-      const binding::Symbol& symbol = context->symbols.symbol(lookup->symbol);
-      if (symbol.kind == binding::SymbolKind::Function) {
-        if (!allow_function_symbol) {
-          return std::unexpected(ResolveDiagnostic{
-              .range = identifier->syntax.range,
-              .message = "This operand does not accept a function symbol.",
-          });
-        }
-        if (auto rejected = reject_shape(checker::OperandShape::Symbol,
-                                         identifier->syntax.range)) {
-          return std::unexpected(std::move(*rejected));
-        }
-        // PTX accepts signed function-address moves with a warning. The
-        // frontend has no warning channel yet, so they remain successful.
-        if (auto rejected =
-                reject_address_type(identifier->syntax.range, true)) {
-          return std::unexpected(std::move(*rejected));
-        }
-        ResolvedFunctionRef function{
-            .spelling = identifier->syntax.text,
-            .symbol_id = symbol.id,
-            .is_entry = symbol.function_is_entry,
-        };
-        if (function.is_entry) {
-          function.address_availability = checker::AvailabilityDescriptor{
-              .minimum_ptx_version = {3, 1},
-              .minimum_sm_version = 35,
-          };
-        }
-        return WithLocs<ResolvedMovSource>{
-            ResolvedMovSource{std::move(function)}, identifier->syntax.range};
-      }
-      is_register = symbol.kind == binding::SymbolKind::Variable &&
-                    symbol.state_space == syntax_ast::AstStateSpace::Register;
-    }
-  }
-
-  if (is_register) {
-    if (auto rejected = reject_shape(checker::OperandShape::Register,
-                                     identifier->syntax.range)) {
-      return std::unexpected(std::move(*rejected));
-    }
-    auto value = resolve_register(operand, context);
-    if (!value)
-      return std::unexpected(value.error());
-    WithLocs<ResolvedMovSource> resolved{
-        ResolvedMovSource{std::move(value->value)}};
-    resolved.locs = std::move(value->locs);
-    return resolved;
-  }
-
-  if (auto rejected = reject_shape(checker::OperandShape::Symbol,
-                                   identifier->syntax.range)) {
-    return std::unexpected(std::move(*rejected));
-  }
-  if (auto rejected = reject_address_type(identifier->syntax.range))
-    return std::unexpected(std::move(*rejected));
-  auto value = resolve_data_symbol(
-      *identifier, context,
-      FormalParameterAddressPolicy::MaterializeDeviceParameter);
-  if (!value)
-    return std::unexpected(value.error());
-  return WithLocs<ResolvedMovSource>{ResolvedMovSource{std::move(*value)},
-                                     identifier->syntax.range};
-}
-
-struct ModifierBindingAttempt {
-  std::optional<ActualModifierTable> modifiers;
-  const syntax_ast::AstModifier* duplicate = nullptr;
-  std::string_view duplicate_slot;
-};
-
-/**
- * Bind source modifier spellings to the slots of one candidate variant.
- *
- * Slot IDs are variant-local. The same spelling may therefore denote `type`
- * in one variant and `result_type` in another, or occupy multiple ordered
- * required/fixed slots in the same variant. The database rejects repeated
- * spellings involving optional slots, so source modifiers greedily consume
- * descriptors in order; optional and absent slots may be skipped, but required
- * slots may not.
- */
-ModifierBindingAttempt bind_modifier_order(
-    const syntax_ast::AstInstruction& ast,
-    const SyntaxVariantDescriptor& variant,
-    std::span<const SyntaxModifierDescriptor> modifiers) {
-  std::unordered_set<std::string_view> slot_ids;
-  for (const auto& descriptor : modifiers) {
-    if (!slot_ids.insert(descriptor.kind_id).second) {
-      throw ResolveException(
-          fmt::format("Variant '{}' contains duplicate modifier slot '{}'.",
-                      variant.variant_name, descriptor.kind_id));
-    }
-  }
-
-  ActualModifierTable result;
-  size_t actual_index = 0;
-  for (const auto& descriptor : modifiers) {
-    if (descriptor.presence == check_end::PresenceRequirement::Absent)
-      continue;
-    if (actual_index < ast.modifiers.size() &&
-        std::ranges::contains(descriptor.allowed_values,
-                              ast.modifiers[actual_index].syntax.text)) {
-      result.emplace(std::string(descriptor.kind_id),
-                     &ast.modifiers[actual_index++]);
-      continue;
-    }
-    if (descriptor.presence == check_end::PresenceRequirement::Required)
-      return {};
-  }
-
-  if (actual_index == ast.modifiers.size())
-    return ModifierBindingAttempt{.modifiers = std::move(result)};
-
-  const auto& extra = ast.modifiers[actual_index];
-  for (auto descriptor = modifiers.rbegin(); descriptor != modifiers.rend();
-       ++descriptor) {
-    if (descriptor->presence != check_end::PresenceRequirement::Absent &&
-        result.contains(std::string(descriptor->kind_id)) &&
-        std::ranges::contains(descriptor->allowed_values, extra.syntax.text)) {
-      return ModifierBindingAttempt{
-          .duplicate = &extra,
-          .duplicate_slot = descriptor->kind_id,
-      };
-    }
-  }
-  return {};
-}
-
-/**
- * Bind against the canonical modifier order and every declared historical
- * order for one semantic variant.
- */
-ModifierBindingAttempt bind_variant_modifiers(
-    const syntax_ast::AstInstruction& ast,
-    const SyntaxVariantDescriptor& variant) {
-  ModifierBindingAttempt result;
-  const auto try_order = [&](std::span<const SyntaxModifierDescriptor> order) {
-    auto attempt = bind_modifier_order(ast, variant, order);
-    if (attempt.modifiers) {
-      if (result.modifiers && *result.modifiers != *attempt.modifiers) {
-        throw ResolveException(fmt::format(
-            "Variant '{}' has modifier orders with ambiguous slot bindings.",
-            variant.variant_name));
-      }
-      if (!result.modifiers)
-        result.modifiers = std::move(attempt.modifiers);
-    } else if (result.duplicate == nullptr && attempt.duplicate != nullptr) {
-      result.duplicate = attempt.duplicate;
-      result.duplicate_slot = attempt.duplicate_slot;
-    }
-  };
-
-  try_order(variant.modifiers);
-  for (const auto& alias : variant.modifier_order_aliases)
-    try_order(alias.modifiers);
-  return result;
-}
-
-bool is_known_modifier_spelling(const SyntaxInstructionDescriptor& instruction,
-                                std::string_view spelling) {
-  return std::ranges::any_of(
-      instruction.variants, [spelling](const auto& variant) {
-        return std::ranges::any_of(
-            variant.modifiers, [spelling](const auto& modifier) {
-              return std::ranges::contains(modifier.allowed_values, spelling);
-            });
-      });
-}
-
-const ResolvedVariantDescriptor& find_resolved_variant_descriptor(
-    const ResolvedInstructionDescriptor& instruction, std::string_view name) {
-  const auto it =
-      std::ranges::find_if(instruction.variants,
-                           [name](const ResolvedVariantDescriptor& descriptor) {
-                             return descriptor.variant_name == name;
-                           });
-  if (it == instruction.variants.end()) {
-    throw ResolveException(
-        fmt::format("Resolved descriptor for '{}' has no variant named '{}'.",
-                    instruction.opcode_name, name));
-  }
-  return *it;
-}
-
-const ResolvedFieldDescriptor& find_resolved_field_descriptor(
-    const ResolvedVariantDescriptor& variant, std::string_view field_id) {
-  const auto it = std::ranges::find_if(
-      variant.fields, [field_id](const ResolvedFieldDescriptor& descriptor) {
-        return descriptor.field_id == field_id;
-      });
-  if (it == variant.fields.end()) {
-    throw ResolveException(
-        fmt::format("Resolved descriptor variant '{}' has no field named '{}'.",
-                    variant.variant_name, field_id));
-  }
-  return *it;
-}
-
-const ResolvedFieldDescriptor& find_resolved_operand_field_descriptor(
-    const ResolvedOperandLayoutDescriptor& layout, std::string_view field_id) {
-  const auto it = std::ranges::find_if(
-      layout.fields, [field_id](const ResolvedFieldDescriptor& descriptor) {
-        return descriptor.field_id == field_id;
-      });
-  if (it == layout.fields.end()) {
-    throw ResolveException(
-        fmt::format("Resolved operand layout '{}' has no field named '{}'.",
-                    layout.layout_id, field_id));
-  }
-  return *it;
-}
-
-const SyntaxModifierDescriptor& find_syntax_modifier_descriptor(
-    const SyntaxVariantDescriptor& variant, std::string_view kind_id) {
-  const auto it = std::ranges::find_if(
-      variant.modifiers, [kind_id](const SyntaxModifierDescriptor& descriptor) {
-        return descriptor.kind_id == kind_id;
-      });
-  if (it == variant.modifiers.end()) {
-    throw ResolveException(
-        fmt::format("Syntax descriptor variant '{}' has no modifier kind '{}'.",
-                    variant.variant_name, kind_id));
-  }
-  return *it;
-}
-
-std::expected<ScalarType, ResolveDiagnostic> type_for_operand(
-    const ResolvedOperandBindingDescriptor& binding,
-    const ResolvedInstructionFields& fields, const SourceRange& range) {
-  const auto& expression = binding.type_expression;
-  if (expression.kind == checker::OperandTypeExpressionKind::None)
-    return ScalarType::Invalid;
-  if (expression.kind == checker::OperandTypeExpressionKind::FixedScalar)
-    return expression.fixed_scalar_type;
-  if (expression.kind != checker::OperandTypeExpressionKind::ModifierField ||
-      expression.modifier_field_id.empty()) {
-    throw ResolveException(
-        fmt::format("Resolved operand field '{}' has an invalid type "
-                    "expression descriptor.",
-                    binding.target_field_id));
-  }
-
-  const std::string_view field_id = expression.modifier_field_id;
-  const auto it = fields.modifiers.find(std::string(field_id));
-  if (it == fields.modifiers.end()) {
-    return std::unexpected(ResolveDiagnostic{
-        .range = range,
-        .message = fmt::format("Operand '{}' requires modifier '{}'.",
-                               binding.target_field_id, field_id),
-    });
-  }
-  if (const auto* type = std::get_if<WithLocs<ScalarType>>(&it->second))
-    return type->value;
-  throw ResolveException(fmt::format(
-      "Operand '{}' expects modifier '{}' to resolve as ScalarType.",
-      binding.target_field_id, field_id));
-}
-
-std::expected<std::optional<uint8_t>, ResolveDiagnostic> vector_arity_for_operand(
-    const ResolvedOperandBindingDescriptor& binding,
-    const ResolvedInstructionFields& fields, const SourceRange& range) {
-  if (binding.vector_arity_modifier_field_id.empty())
-    return std::nullopt;
-
-  const std::string_view field_id = binding.vector_arity_modifier_field_id;
-  const auto it = fields.modifiers.find(std::string(field_id));
-  if (it == fields.modifiers.end()) {
-    return std::unexpected(ResolveDiagnostic{
-        .range = range,
-        .message = fmt::format("Operand '{}' requires vector arity field '{}'.",
-                               binding.target_field_id, field_id),
-    });
-  }
-  if (const auto* arity = std::get_if<WithLocs<VectorArity>>(&it->second))
-    return vector_arity_count(arity->value);
-  throw ResolveException(fmt::format(
-      "Operand '{}' expects modifier '{}' to resolve as VectorArity.",
-      binding.target_field_id, field_id));
-}
-
-std::expected<ResolvedFieldValue, ResolveDiagnostic> resolve_operand_value(
-    const ResolvedFieldDescriptor& field,
-    const ResolvedOperandBindingDescriptor& binding,
-    const syntax_ast::AstOperand& operand,
-    const ResolvedInstructionFields& fields, const ResolveContext* context) {
-  switch (field.value_kind) {
-    case ResolvedValueKind::Register: {
-      auto value = resolve_register(operand, context);
-      if (!value)
-        return std::unexpected(value.error());
-      return ResolvedFieldValue{std::move(*value)};
-    }
-    case ResolvedValueKind::MbarrierStateToken: {
-      const bool is_sink = std::get_if<syntax_ast::AstIdentifierRef>(&operand) != nullptr &&
-                           std::get<syntax_ast::AstIdentifierRef>(operand).syntax.text == "_";
-      if (is_sink) {
-        if (binding.mbarrier_state_token_form ==
-            checker::MbarrierStateTokenForm::Register) {
-          return std::unexpected(ResolveDiagnostic{.range = syntax_ast::sourceRange(operand),
-              .message = "The '_' sink is not allowed for this mbarrier state token."});
-        }
-        return ResolvedFieldValue{WithLocs<ResolvedMbarrierStateToken>{
-            ResolvedMbarrierStateToken{.register_ref = std::nullopt},
-            syntax_ast::sourceRange(operand)}};
-      }
-      if (binding.mbarrier_state_token_form == checker::MbarrierStateTokenForm::Sink) {
-        return std::unexpected(ResolveDiagnostic{.range = syntax_ast::sourceRange(operand),
-            .message = "This mbarrier state token requires the '_' sink."});
-      }
-      auto value = resolve_register(operand, context);
-      if (!value)
-        return std::unexpected(value.error());
-      WithLocs<ResolvedMbarrierStateToken> token{
-          ResolvedMbarrierStateToken{.register_ref = std::move(value->value)}};
-      token.locs = std::move(value->locs);
-      return ResolvedFieldValue{std::move(token)};
-    }
-    case ResolvedValueKind::RegisterOrSink: {
-      if (const auto* identifier =
-              std::get_if<syntax_ast::AstIdentifierRef>(&operand);
-          identifier != nullptr && identifier->syntax.text == "_") {
-        return ResolvedFieldValue{WithLocs<ResolvedRegisterOrSink>{
-            ResolvedRegisterOrSink{.register_ref = std::nullopt},
-            identifier->syntax.range}};
-      }
-      auto value = resolve_register(operand, context);
-      if (!value)
-        return std::unexpected(value.error());
-      WithLocs<ResolvedRegisterOrSink> destination{
-          ResolvedRegisterOrSink{.register_ref = std::move(value->value)}};
-      destination.locs = std::move(value->locs);
-      return ResolvedFieldValue{std::move(destination)};
-    }
-    case ResolvedValueKind::Predicate: {
-      auto value = resolve_predicate(operand, context);
-      if (!value)
-        return std::unexpected(value.error());
-      return ResolvedFieldValue{std::move(*value)};
-    }
-    case ResolvedValueKind::PredicateSource: {
-      auto value = resolve_predicate_source(operand, context);
-      if (!value)
-        return std::unexpected(value.error());
-      return ResolvedFieldValue{std::move(*value)};
-    }
-    case ResolvedValueKind::Immediate: {
-      const auto* immediate = std::get_if<syntax_ast::AstImmediate>(&operand);
-      if (immediate == nullptr) {
-        return std::unexpected(ResolveDiagnostic{
-            .range = syntax_ast::sourceRange(operand),
-            .message = "Expected an immediate operand.",
-        });
-      }
-      const auto type =
-          type_for_operand(binding, fields, immediate->syntax.range);
-      if (!type)
-        return std::unexpected(type.error());
-      auto value = resolve_immediate_value(
-          *immediate, *type,
-          binding.immediate_conversion_policy ==
-              checker::ImmediateConversionPolicy::RequireTargetRange);
-      if (!value)
-        return std::unexpected(value.error());
-      return ResolvedFieldValue{WithLocs<ResolvedImmediate>{
-          std::move(*value), immediate->syntax.range}};
-    }
-    case ResolvedValueKind::RegOrImm: {
-      const auto type =
-          type_for_operand(binding, fields, syntax_ast::sourceRange(operand));
-      if (!type)
-        return std::unexpected(type.error());
-      auto value = resolve_reg_or_imm(
-          operand, *type, context,
-          binding.immediate_conversion_policy ==
-              checker::ImmediateConversionPolicy::RequireTargetRange);
-      if (!value)
-        return std::unexpected(value.error());
-      return ResolvedFieldValue{std::move(*value)};
-    }
-    case ResolvedValueKind::ShflDestination: {
-      auto value = resolve_shfl_destination(
-          operand, binding.allow_destination_sink,
-          binding.allow_predicate_sink, context);
-      if (!value)
-        return std::unexpected(value.error());
-      return ResolvedFieldValue{std::move(*value)};
-    }
-    case ResolvedValueKind::PredicatePair: {
-      auto value = resolve_predicate_pair(operand, context);
-      if (!value)
-        return std::unexpected(value.error());
-      return ResolvedFieldValue{std::move(*value)};
-    }
-    case ResolvedValueKind::MovSource: {
-      const auto type =
-          type_for_operand(binding, fields, syntax_ast::sourceRange(operand));
-      if (!type)
-        return std::unexpected(type.error());
-      auto value =
-          resolve_mov_source(operand, *type, binding.allowed_shapes,
-                             binding.allow_function_symbol, context);
-      if (!value)
-        return std::unexpected(value.error());
-      return ResolvedFieldValue{std::move(*value)};
-    }
-    case ResolvedValueKind::VectorRegister: {
-      auto value = resolve_vector_register(operand, context);
-      if (!value)
-        return std::unexpected(value.error());
-      return ResolvedFieldValue{std::move(*value)};
-    }
-    case ResolvedValueKind::VectorSpecialRegister: {
-      auto value = resolve_vector_special_register(operand);
-      if (!value)
-        return std::unexpected(value.error());
-      return ResolvedFieldValue{std::move(*value)};
-    }
-    case ResolvedValueKind::BranchTarget: {
-      auto value = resolve_branch_target(operand, context);
-      if (!value)
-        return std::unexpected(value.error());
-      return ResolvedFieldValue{std::move(*value)};
-    }
-    case ResolvedValueKind::BranchTargetSet: {
-      auto value = resolve_branch_target_set(operand, context);
-      if (!value)
-        return std::unexpected(value.error());
-      return ResolvedFieldValue{std::move(*value)};
-    }
-    case ResolvedValueKind::SpecialRegister: {
-      auto value = resolve_special_register(operand);
-      if (!value)
-        return std::unexpected(value.error());
-      return ResolvedFieldValue{std::move(*value)};
-    }
-    case ResolvedValueKind::Symbol: {
-      auto value = resolve_symbol(operand, context);
-      if (!value)
-        return std::unexpected(value.error());
-      return ResolvedFieldValue{std::move(*value)};
-    }
-    case ResolvedValueKind::Address: {
-      auto value = resolve_address(operand, context);
-      if (!value)
-        return std::unexpected(value.error());
-      return ResolvedFieldValue{std::move(*value)};
-    }
-    case ResolvedValueKind::RegisterVector: {
-      if (binding.minimum_elements != 0) {
-        auto value = resolve_modern_register_vector(operand, binding, context);
-        if (!value)
-          return std::unexpected(value.error());
-        return ResolvedFieldValue{std::move(*value)};
-      }
-      const auto type =
-          type_for_operand(binding, fields, syntax_ast::sourceRange(operand));
-      if (!type)
-        return std::unexpected(type.error());
-      const auto arity =
-          vector_arity_for_operand(binding, fields, syntax_ast::sourceRange(operand));
-      if (!arity)
-        return std::unexpected(arity.error());
-      auto value =
-          resolve_reg_vector(operand, *type,
-                             binding.allowed_vector_arities,
-                             *arity,
-                             binding.vector_type_policy,
-                             binding.register_width_policy,
-                             binding.allow_vector_sink,
-                             binding.vector_sink_payload_bits, context);
-      if (!value)
-        return std::unexpected(value.error());
-      return ResolvedFieldValue{std::move(*value)};
-    }
-    case ResolvedValueKind::TensorCoordinate: {
-      auto value = resolve_tensor_coordinate(operand, binding, fields, context);
-      if (!value)
-        return std::unexpected(value.error());
-      return ResolvedFieldValue{std::move(*value)};
-    }
-    case ResolvedValueKind::DirectCallTarget: {
-      auto value = resolve_direct_call_target(operand, context);
-      if (!value)
-        return std::unexpected(value.error());
-      return ResolvedFieldValue{std::move(*value)};
-    }
-    case ResolvedValueKind::IndirectCallee: {
-      auto value = resolve_indirect_callee(operand, context);
-      if (!value)
-        return std::unexpected(value.error());
-      return ResolvedFieldValue{std::move(*value)};
-    }
-    case ResolvedValueKind::CallReturnParameter: {
-      auto value = resolve_call_return_parameter(operand, context);
-      if (!value)
-        return std::unexpected(value.error());
-      return ResolvedFieldValue{std::move(*value)};
-    }
-    case ResolvedValueKind::CallArguments: {
-      auto value = resolve_call_arguments(operand, context);
-      if (!value)
-        return std::unexpected(value.error());
-      return ResolvedFieldValue{std::move(*value)};
-    }
-    case ResolvedValueKind::Bool:
-    case ResolvedValueKind::ScalarType:
-    case ResolvedValueKind::RoundingMode:
-    case ResolvedValueKind::ComparisonOperator:
-    case ResolvedValueKind::BooleanOperator:
-    case ResolvedValueKind::CacheOperator:
-    case ResolvedValueKind::EvictionPriority:
-    case ResolvedValueKind::MemoryConsistency:
-    case ResolvedValueKind::MemoryScope:
-    case ResolvedValueKind::VectorArity:
-    case ResolvedValueKind::MemoryStateSpace:
-    case ResolvedValueKind::MbarrierPhaseType:
-    case ResolvedValueKind::MbarrierLayout:
-    case ResolvedValueKind::AsyncProxyKind:
-    case ResolvedValueKind::ProxyKindPair:
-      throw ResolveException(fmt::format(
-          "Operand slot '{}' has a non-operand resolved value kind.",
-          field.field_id));
-  }
-  throw ResolveException("Unknown ResolvedValueKind.");
-}
-
-ResolvedFieldValue resolve_default_modifier_value(
-    const ResolvedFieldDescriptor& field,
-    const ResolvedModifierBindingDescriptor& binding) {
-  const auto& default_value = binding.default_value;
-  switch (field.value_kind) {
-    case ResolvedValueKind::Bool:
-      if (default_value.kind != ResolvedModifierDefaultKind::Bool) {
-        throw ResolveException(fmt::format(
-            "Optional modifier '{}' requires a boolean default for resolved "
-            "field '{}'.",
-            binding.source_kind_id, field.field_id));
-      }
-      return ResolvedFieldValue{WithLocs<bool>{default_value.bool_value}};
-    case ResolvedValueKind::ScalarType:
-      if (default_value.kind != ResolvedModifierDefaultKind::ScalarType ||
-          default_value.scalar_type == ScalarType::Invalid) {
-        throw ResolveException(fmt::format(
-            "Optional modifier '{}' requires a scalar-type default for "
-            "resolved field '{}'.",
-            binding.source_kind_id, field.field_id));
-      }
-      return ResolvedFieldValue{
-          WithLocs<ScalarType>{default_value.scalar_type}};
-    case ResolvedValueKind::RoundingMode:
-      if (default_value.kind != ResolvedModifierDefaultKind::RoundingMode ||
-          default_value.rounding_mode == RoundingMode::Invalid) {
-        throw ResolveException(fmt::format(
-            "Optional modifier '{}' requires a rounding-mode default for "
-            "resolved field '{}'.",
-            binding.source_kind_id, field.field_id));
-      }
-      return ResolvedFieldValue{
-          WithLocs<RoundingMode>{default_value.rounding_mode}};
-    case ResolvedValueKind::ComparisonOperator:
-      throw ResolveException(fmt::format(
-          "Optional modifier '{}' cannot use a comparison-operator default "
-          "for resolved field '{}'.",
-          binding.source_kind_id, field.field_id));
-    case ResolvedValueKind::BooleanOperator:
-      throw ResolveException(fmt::format(
-          "Optional modifier '{}' cannot use a boolean-operator default for "
-          "resolved field '{}'.",
-          binding.source_kind_id, field.field_id));
-    case ResolvedValueKind::CacheOperator:
-      if (default_value.kind != ResolvedModifierDefaultKind::CacheOperator) {
-        throw ResolveException(fmt::format(
-            "Optional modifier '{}' requires a cache-operator default for "
-            "resolved field '{}'.",
-            binding.source_kind_id, field.field_id));
-      }
-      return ResolvedFieldValue{WithLocs<CacheOperator>{
-          default_value.cache_operator}};
-    case ResolvedValueKind::EvictionPriority:
-      throw ResolveException(fmt::format(
-          "Optional modifier '{}' cannot use an eviction-priority default "
-          "for resolved field '{}'.",
-          binding.source_kind_id, field.field_id));
-    case ResolvedValueKind::MemoryConsistency:
-      if (default_value.kind != ResolvedModifierDefaultKind::MemoryConsistency) {
-        throw ResolveException(fmt::format(
-            "Optional modifier '{}' requires a memory-consistency default for "
-            "resolved field '{}'.", binding.source_kind_id, field.field_id));
-      }
-      return ResolvedFieldValue{WithLocs<MemoryConsistency>{
-          default_value.memory_consistency}};
-    case ResolvedValueKind::MemoryScope:
-      if (default_value.kind != ResolvedModifierDefaultKind::MemoryScope) {
-        throw ResolveException(fmt::format(
-            "Optional modifier '{}' requires a memory-scope default for "
-            "resolved field '{}'.", binding.source_kind_id, field.field_id));
-      }
-      return ResolvedFieldValue{WithLocs<MemoryScope>{default_value.memory_scope}};
-    case ResolvedValueKind::MemoryStateSpace:
-      if (default_value.kind !=
-              ResolvedModifierDefaultKind::MemoryStateSpace ||
-          default_value.memory_state_space == MemoryStateSpace::Invalid) {
-        throw ResolveException(fmt::format(
-            "Optional modifier '{}' requires a memory-state-space default "
-            "for resolved field '{}'.",
-            binding.source_kind_id, field.field_id));
-      }
-      return ResolvedFieldValue{WithLocs<MemoryStateSpace>{
-          default_value.memory_state_space}};
-    case ResolvedValueKind::MbarrierPhaseType:
-      if (default_value.kind !=
-          ResolvedModifierDefaultKind::MbarrierPhaseType) {
-        throw ResolveException(fmt::format(
-            "Optional modifier '{}' requires an mbarrier phase-type default "
-            "for resolved field '{}'.", binding.source_kind_id, field.field_id));
-      }
-      return ResolvedFieldValue{WithLocs<MbarrierPhaseType>{
-          default_value.mbarrier_phase_type}};
-    case ResolvedValueKind::MbarrierLayout:
-      if (default_value.kind != ResolvedModifierDefaultKind::MbarrierLayout) {
-        throw ResolveException(fmt::format(
-            "Optional modifier '{}' requires an mbarrier layout default for "
-            "resolved field '{}'.", binding.source_kind_id, field.field_id));
-      }
-      return ResolvedFieldValue{WithLocs<MbarrierLayout>{
-          default_value.mbarrier_layout}};
-    case ResolvedValueKind::AsyncProxyKind:
-      if (default_value.kind != ResolvedModifierDefaultKind::AsyncProxyKind) {
-        throw ResolveException(fmt::format(
-            "Optional modifier '{}' requires an async proxy default for "
-            "resolved field '{}'.", binding.source_kind_id, field.field_id));
-      }
-      return ResolvedFieldValue{WithLocs<AsyncProxyKind>{
-          default_value.async_proxy_kind}};
-    case ResolvedValueKind::ProxyKindPair:
-      if (default_value.kind != ResolvedModifierDefaultKind::ProxyKindPair) {
-        throw ResolveException(fmt::format(
-            "Optional modifier '{}' requires a proxy pair default for "
-            "resolved field '{}'.", binding.source_kind_id, field.field_id));
-      }
-      return ResolvedFieldValue{WithLocs<ProxyKindPair>{
-          default_value.proxy_kind_pair}};
-    case ResolvedValueKind::Register:
-    case ResolvedValueKind::Predicate:
-    case ResolvedValueKind::PredicateSource:
-    case ResolvedValueKind::Immediate:
-    case ResolvedValueKind::RegOrImm:
-    case ResolvedValueKind::ShflDestination:
-    case ResolvedValueKind::PredicatePair:
-    case ResolvedValueKind::MovSource:
-    case ResolvedValueKind::VectorRegister:
-    case ResolvedValueKind::VectorSpecialRegister:
-    case ResolvedValueKind::BranchTarget:
-    case ResolvedValueKind::BranchTargetSet:
-    case ResolvedValueKind::SpecialRegister:
-    case ResolvedValueKind::Symbol:
-    case ResolvedValueKind::Address:
-    case ResolvedValueKind::MbarrierStateToken:
-    case ResolvedValueKind::RegisterOrSink:
-    case ResolvedValueKind::RegisterVector:
-    case ResolvedValueKind::TensorCoordinate:
-    case ResolvedValueKind::DirectCallTarget:
-    case ResolvedValueKind::IndirectCallee:
-    case ResolvedValueKind::CallReturnParameter:
-    case ResolvedValueKind::CallArguments:
-    case ResolvedValueKind::VectorArity:
-      throw ResolveException(fmt::format(
-          "Optional modifier '{}' targets non-modifier resolved field '{}'.",
-          binding.source_kind_id, field.field_id));
-  }
-  throw ResolveException("Unknown ResolvedValueKind.");
-}
-
-}  // namespace
-
-checker::AvailabilityDescriptor special_register_availability(
-    const base::Info& info) {
-  checker::AvailabilityDescriptor availability{
-      .minimum_ptx_version = {info.minimum_ptx_major, info.minimum_ptx_minor},
-      .minimum_sm_version = info.minimum_sm,
-  };
-  if (info.required_capability.empty())
-    return availability;
-
-  availability.any_of[0] = checker::AvailabilityClause{
-      .minimum_ptx_version = {info.minimum_ptx_major, info.minimum_ptx_minor},
-      .minimum_sm_version = info.minimum_sm,
-      .capabilities = {info.required_capability},
-      .capability_count = 1,
-  };
-  availability.any_of_count = 1;
-  return availability;
-}
-
-std::expected<ResolvedImmediate, ResolveDiagnostic> resolve_immediate_literal(
-    const syntax_ast::AstImmediate& immediate, ScalarType type) {
-  return resolve_immediate_value(immediate, type);
-}
-
-std::expected<WithLocs<ResolvedImmediate>, ResolveDiagnostic>
-resolve_call_literal(
-    const ResolvedCallLiteral& literal, SourceRange range,
-    const declaration_semantics::FunctionParameterContract& formal) {
-  if (formal.scalar_type == ScalarType::Invalid) {
-    return std::unexpected(ResolveDiagnostic{
-        .range = range,
-        .message = fmt::format(
-            "Call literal '{}' has unsupported formal scalar type '{}'.",
-            literal.spelling, formal.type_spelling),
-    });
-  }
-  const syntax_ast::AstImmediate immediate{
-      .syntax = {.text = literal.spelling, .range = range},
-      .kind = literal.kind,
-  };
-  // Call arguments retain their formal-parameter representability contract.
-  auto resolved = resolve_immediate_value(immediate, formal.scalar_type, true);
-  if (!resolved)
-    return std::unexpected(resolved.error());
-  return WithLocs<ResolvedImmediate>{std::move(*resolved), range};
-}
-
-std::expected<ActualModifierTable, ResolveDiagnostic> collect_actual_modifiers(
-    const syntax_ast::AstInstruction& ast,
-    const check_end::SyntaxVariantDescriptor& variant) {
-  auto attempt = bind_variant_modifiers(ast, variant);
-  if (attempt.modifiers)
-    return std::move(*attempt.modifiers);
-  if (attempt.duplicate != nullptr) {
-    return std::unexpected(ResolveDiagnostic{
-        .range = attempt.duplicate->syntax.range,
-        .message =
-            fmt::format("Duplicate '{}' modifier.", attempt.duplicate_slot),
-    });
-  }
-  return std::unexpected(ResolveDiagnostic{
-      .range = ast.range,
-      .message = fmt::format(
-          "Modifier combination does not match instruction variant '{}'.",
-          variant.variant_name),
-  });
-}
-
-std::expected<std::string_view, ResolveDiagnostic> select_variant_name(
-    const syntax_ast::AstInstruction& ast,
-    const check_end::SyntaxInstructionDescriptor& instruction) {
-  if (ast.opcode.syntax.text != instruction.Opcode_name) {
-    return std::unexpected(ResolveDiagnostic{
-        .range = ast.opcode.syntax.range,
-        .message = fmt::format("Cannot resolve opcode '{}' as '{}'.",
-                               ast.opcode.syntax.text, instruction.Opcode_name),
-    });
-  }
-
-  for (const auto& modifier : ast.modifiers) {
-    if (!is_known_modifier_spelling(instruction, modifier.syntax.text)) {
-      return std::unexpected(ResolveDiagnostic{
-          .range = modifier.syntax.range,
-          .message =
-              fmt::format("Unknown modifier '{}'.", modifier.syntax.text),
-      });
-    }
-  }
-
-  std::optional<std::string_view> selected;
-  const syntax_ast::AstModifier* duplicate = nullptr;
-  std::string_view duplicate_slot;
-  for (const auto& variant : instruction.variants) {
-    auto attempt = bind_variant_modifiers(ast, variant);
-    if (!attempt.modifiers) {
-      if (duplicate == nullptr && attempt.duplicate != nullptr) {
-        duplicate = attempt.duplicate;
-        duplicate_slot = attempt.duplicate_slot;
-      }
-      continue;
-    }
-
-    if (selected) {
-      return std::unexpected(ResolveDiagnostic{
-          .range = ast.range,
-          .message = fmt::format(
-              "Ambiguous modifier combination for instruction '{}'.",
-              ast.opcode.syntax.text),
-      });
-    }
-    selected = variant.variant_name;
-  }
-
-  if (selected)
-    return *selected;
-  if (duplicate != nullptr) {
-    return std::unexpected(ResolveDiagnostic{
-        .range = duplicate->syntax.range,
-        .message = fmt::format("Duplicate '{}' modifier.", duplicate_slot),
-    });
-  }
-  return std::unexpected(ResolveDiagnostic{
-      .range = ast.range,
-      .message = fmt::format(
-          "No variant of instruction '{}' accepts this modifier combination.",
-          ast.opcode.syntax.text),
-  });
-}
-
 std::expected<ResolvedInstructionFields, ResolveDiagnostic> resolve_fields(
     const syntax_ast::AstInstruction& ast,
     const check_end::SyntaxInstructionDescriptor& syntax_instruction,
     const check_end::ResolvedInstructionDescriptor& resolved_instruction,
     std::string_view variant_name, const ResolveContext* context) {
   const SyntaxVariantDescriptor& syntax_variant =
-      find_syntax_variant_descriptor(syntax_instruction, variant_name);
+      detail::find_syntax_variant_descriptor(syntax_instruction, variant_name);
   const ResolvedVariantDescriptor& resolved_variant =
-      find_resolved_variant_descriptor(resolved_instruction, variant_name);
-  const auto selected_layout = select_operand_layout(syntax_variant, ast);
+      detail::find_resolved_variant_descriptor(resolved_instruction, variant_name);
+  const auto selected_layout = detail::select_operand_layout(syntax_variant, ast);
   if (!selected_layout)
     return std::unexpected(selected_layout.error());
 
@@ -3222,7 +97,7 @@ std::expected<ResolvedInstructionFields, ResolveDiagnostic> resolve_fields(
           selected_layout->index)},
   };
   if (ast.predicate) {
-    auto predicate = resolve_predicate_identifier(
+    auto predicate = detail::resolve_predicate_identifier(
         ast.predicate->name, ast.predicate->negated, ast.predicate->range,
         context);
     if (!predicate)
@@ -3231,9 +106,9 @@ std::expected<ResolvedInstructionFields, ResolveDiagnostic> resolve_fields(
   }
   for (const auto& binding : resolved_variant.modifier_bindings) {
     const auto& syntax_modifier =
-        find_syntax_modifier_descriptor(syntax_variant, binding.source_kind_id);
-    const auto& field = find_resolved_field_descriptor(resolved_variant,
-                                                       binding.target_field_id);
+        detail::find_syntax_modifier_descriptor(syntax_variant, binding.source_kind_id);
+    const auto& field = detail::find_resolved_field_descriptor(resolved_variant,
+                                                               binding.target_field_id);
     const auto actual =
         actual_modifiers->find(std::string(binding.source_kind_id));
     const bool present = actual != actual_modifiers->end();
@@ -3248,7 +123,8 @@ std::expected<ResolvedInstructionFields, ResolveDiagnostic> resolve_fields(
       if (syntax_modifier.presence ==
           check_end::PresenceRequirement::Optional) {
         fields.modifiers.emplace(
-            field.field_id, resolve_default_modifier_value(field, binding));
+            field.field_id,
+            detail::resolve_default_modifier_value(field, binding));
         continue;
       }
       return std::unexpected(ResolveDiagnostic{
@@ -3258,136 +134,25 @@ std::expected<ResolvedInstructionFields, ResolveDiagnostic> resolve_fields(
       });
     }
 
-    switch (field.value_kind) {
-      case ResolvedValueKind::Bool:
-        fields.modifiers.emplace(
-            field.field_id, WithLocs<bool>{true, actual->second->syntax.range});
-        break;
-      case ResolvedValueKind::ScalarType: {
-        auto value = resolve_scalar_type(*actual->second);
-        if (!value)
-          return std::unexpected(value.error());
-        fields.modifiers.emplace(field.field_id, std::move(*value));
-      } break;
-      case ResolvedValueKind::RoundingMode: {
-        auto value = resolve_rounding_mode(*actual->second);
-        if (!value)
-          return std::unexpected(value.error());
-        fields.modifiers.emplace(field.field_id, std::move(*value));
-      } break;
-      case ResolvedValueKind::ComparisonOperator: {
-        auto value = resolve_comparison_operator(*actual->second);
-        if (!value)
-          return std::unexpected(value.error());
-        fields.modifiers.emplace(field.field_id, std::move(*value));
-      } break;
-      case ResolvedValueKind::BooleanOperator: {
-        auto value = resolve_boolean_operator(*actual->second);
-        if (!value)
-          return std::unexpected(value.error());
-        fields.modifiers.emplace(field.field_id, std::move(*value));
-      } break;
-      case ResolvedValueKind::CacheOperator: {
-        auto value = resolve_cache_operator(*actual->second);
-        if (!value)
-          return std::unexpected(value.error());
-        fields.modifiers.emplace(field.field_id, std::move(*value));
-      } break;
-      case ResolvedValueKind::EvictionPriority: {
-        auto value = resolve_eviction_priority(*actual->second);
-        if (!value)
-          return std::unexpected(value.error());
-        fields.modifiers.emplace(field.field_id, std::move(*value));
-      } break;
-      case ResolvedValueKind::MemoryConsistency: {
-        auto value = resolve_memory_consistency(*actual->second);
-        if (!value)
-          return std::unexpected(value.error());
-        fields.modifiers.emplace(field.field_id, std::move(*value));
-      } break;
-      case ResolvedValueKind::MemoryScope: {
-        auto value = resolve_memory_scope(*actual->second);
-        if (!value)
-          return std::unexpected(value.error());
-        fields.modifiers.emplace(field.field_id, std::move(*value));
-      } break;
-      case ResolvedValueKind::VectorArity: {
-        auto value = resolve_vector_arity(*actual->second);
-        if (!value)
-          return std::unexpected(value.error());
-        fields.modifiers.emplace(field.field_id, std::move(*value));
-      } break;
-      case ResolvedValueKind::MemoryStateSpace: {
-        auto value = resolve_memory_state_space(*actual->second);
-        if (!value)
-          return std::unexpected(value.error());
-        fields.modifiers.emplace(field.field_id, std::move(*value));
-      } break;
-      case ResolvedValueKind::MbarrierPhaseType: {
-        auto value = resolve_mbarrier_phase_type(*actual->second);
-        if (!value)
-          return std::unexpected(value.error());
-        fields.modifiers.emplace(field.field_id, std::move(*value));
-      } break;
-      case ResolvedValueKind::MbarrierLayout: {
-        auto value = resolve_mbarrier_layout(*actual->second);
-        if (!value)
-          return std::unexpected(value.error());
-        fields.modifiers.emplace(field.field_id, std::move(*value));
-      } break;
-      case ResolvedValueKind::AsyncProxyKind: {
-        auto value = resolve_async_proxy_kind(*actual->second);
-        if (!value)
-          return std::unexpected(value.error());
-        fields.modifiers.emplace(field.field_id, std::move(*value));
-      } break;
-      case ResolvedValueKind::ProxyKindPair: {
-        auto value = resolve_proxy_kind_pair(*actual->second);
-        if (!value)
-          return std::unexpected(value.error());
-        fields.modifiers.emplace(field.field_id, std::move(*value));
-      } break;
-      case ResolvedValueKind::Register:
-      case ResolvedValueKind::Predicate:
-      case ResolvedValueKind::PredicateSource:
-      case ResolvedValueKind::Immediate:
-      case ResolvedValueKind::RegOrImm:
-      case ResolvedValueKind::ShflDestination:
-      case ResolvedValueKind::PredicatePair:
-      case ResolvedValueKind::MovSource:
-      case ResolvedValueKind::VectorRegister:
-      case ResolvedValueKind::VectorSpecialRegister:
-      case ResolvedValueKind::BranchTarget:
-      case ResolvedValueKind::BranchTargetSet:
-      case ResolvedValueKind::SpecialRegister:
-      case ResolvedValueKind::Symbol:
-      case ResolvedValueKind::Address:
-      case ResolvedValueKind::RegisterVector:
-      case ResolvedValueKind::DirectCallTarget:
-      case ResolvedValueKind::IndirectCallee:
-      case ResolvedValueKind::CallReturnParameter:
-      case ResolvedValueKind::CallArguments:
-      case ResolvedValueKind::MbarrierStateToken:
-      case ResolvedValueKind::RegisterOrSink:
-        throw ResolveException(
-            fmt::format("Modifier '{}' has a non-modifier resolved value kind.",
-                        binding.source_kind_id));
-    }
+    auto value = detail::resolve_modifier_value(field, *actual->second);
+    if (!value)
+      return std::unexpected(value.error());
+    fields.modifiers.emplace(field.field_id, std::move(*value));
   }
 
   for (size_t index = 0; index < ast.operands.size(); ++index) {
     const auto& binding = resolved_layout.bindings[index];
-    const auto& field = find_resolved_operand_field_descriptor(
+    const auto& field = detail::find_resolved_operand_field_descriptor(
         resolved_layout, binding.target_field_id);
-    auto value = resolve_operand_value(field, binding, ast.operands[index],
-                                       fields, context);
+    auto value = detail::resolve_operand_value(field, binding, ast.operands[index],
+                                               fields, context);
     if (!value)
       return std::unexpected(value.error());
     if (auto* address = std::get_if<WithLocs<ResolvedAddress>>(&*value)) {
       const auto state_space = actual_modifiers->find("state_space");
       if (state_space != actual_modifiers->end()) {
         address->value.parameter_qualifier =
-            parameter_address_qualifier_from_modifier(
+            detail::parameter_address_qualifier_from_modifier(
                 state_space->second->syntax.text);
       }
     }
@@ -3401,48 +166,6 @@ std::expected<ResolvedInstructionFields, ResolveDiagnostic> resolve_fields(
   }
 
   return fields;
-}
-
-check_end::OperandSyntaxShape check_end::get_operand_syntax_shape(
-    const syntax_ast::AstOperand& operand) {
-  return std::visit(
-      [](const auto& item) -> check_end::OperandSyntaxShape {
-        using Item = std::remove_cvref_t<decltype(item)>;
-        if constexpr (std::same_as<Item, syntax_ast::AstIdentifierRef>) {
-          return check_end::OperandSyntaxShape::Identifier;
-        } else if constexpr (std::same_as<Item, syntax_ast::AstImmediate>) {
-          return check_end::OperandSyntaxShape::Immediate;
-
-        } else if constexpr (std::same_as<Item,
-                                          syntax_ast::AstPredicateOperand>) {
-          return check_end::OperandSyntaxShape::Predicate;
-
-        } else if constexpr (std::same_as<Item, syntax_ast::AstAddress>) {
-          return check_end::OperandSyntaxShape::Address;
-
-        } else if constexpr (std::same_as<Item, syntax_ast::AstVectorPack>) {
-          return check_end::OperandSyntaxShape::VectorPack;
-
-        } else if constexpr (std::same_as<Item, syntax_ast::AstVectorMember>) {
-          return check_end::OperandSyntaxShape::VectorMember;
-        } else if constexpr (std::same_as<Item,
-                                          syntax_ast::AstCallParameterList>) {
-          return check_end::OperandSyntaxShape::Group;
-        } else if constexpr (std::same_as<Item, syntax_ast::AstCallTarget>) {
-          return check_end::OperandSyntaxShape::CallTarget;
-        } else if constexpr (std::same_as<Item, syntax_ast::AstCallTargetSet>) {
-          return check_end::OperandSyntaxShape::CallTargetSet;
-        } else if constexpr (std::same_as<Item,
-                                          syntax_ast::AstBranchTargetSet>) {
-          return check_end::OperandSyntaxShape::BranchTargetSet;
-        } else if constexpr (std::same_as<Item,
-                                          syntax_ast::AstRegisterPredicatePair>) {
-          return check_end::OperandSyntaxShape::RegisterPredicatePair;
-        } else {
-          return check_end::OperandSyntaxShape::BranchTarget;
-        }
-      },
-      operand);
 }
 
 };  // namespace ptx_frontend::resolved_ir
